### 2022-11-14 #4928 @editorialbot
[REVIEW]: DTCC Builder: A mesh generator for automatic, efficient, and robust mesh generation for large-scale city modeling and simulation
**Submitting author:** @alogg (<a href="http://orcid.org/0000-0002-1547-4773">Anders Logg</a>)
**Repository:** https://github.com/dtcc-platform/dtcc-builder
**Branch with paper.md** (empty if default branch): joss-2022
**Version:** v0.9
