### 2022-11-15 4d199a6
Add Protobuf output for Surface3D and write CitySurface.pb from dtcc-build
