+++ docker/data-sync.yml
@@ -2,5 +2,7 @@ version: "2"
 syncs:
   data_sync:
     sync_strategy: 'native_osx'
-    src: '../' #path to the volume you want to synchronise
+    src: '..' #path to the volume you want to synchronise
     host_disk_mount_mode: 'cached'
+    watch_excludes: ['.*/.git', '.*/node_modules', '.gitignore']
+    sync_excludes: ['Gemfile.lock', 'Gemfile', 'config.rb', 'package.json', 'Gruntfile*', 'bower_components', 'node_modules', '.git', '*.coffee', '*.scss', '*.sass']
