### 2022-11-15 f1ceaa5
export building ground height
