+++ docker/docker-sync.yml
@@ -1,0 +1,9 @@
+version: "2"
+syncs:
+  data_sync:
+    sync_strategy: 'native_osx'
+    src: '../' #path to the volume you want to synchronise
+    host_disk_mount_mode: 'cached'
+    sync_userid: ${_UID}
+    sync_groupid: ${GID}
+    monit_enable: true
