+++ docker/data-sync.yml
@@ -1,10 +1,0 @@
-version: "2"
-syncs:
-  data_sync:
-    sync_strategy: 'native_osx'
-    src: '..' #path to the volume you want to synchronise
-    host_disk_mount_mode: 'cached'
-    watch_excludes: ['.*/.git', '.*/node_modules', '.gitignore']
-    sync_excludes: ['Gemfile.lock', 'Gemfile', 'config.rb', 'package.json', 'Gruntfile*', 'bower_components', 'node_modules', '.git', '*.coffee', '*.scss', '*.sass']
-    sync_userid: ${_UID}
-    sync_groupid: ${GID}
