### 2022-11-17 2bfeef1
Updated .gitlab-ci.yml file
