### 2022-11-28 9ad52d8
remove a bunch of debug output
