+++ include/BoundingBoxTree.h
@@ -46,8 +46,8 @@ public:
   // Build bounding box tree for objects (defined by their bounding boxes)
   void Build(const std::vector<BoundingBox2D> &bboxes)
   {
-    info("BoundingBoxTree: Building 2D bounding box tree for " +
-         str(bboxes.size()) + " objects...");
+    // info("BoundingBoxTree: Building 2D bounding box tree for " +
+    //     str(bboxes.size()) + " objects...");
 
     // Clear tree if built before
     Nodes.clear();
