### 2022-12-01 46a4c7b
Add mesh smoothing evaluation to sandbox
