+++ sandbox/smoothing-2022/test.py
@@ -1,0 +1,290 @@
+# Simple test to compare simple (ad hoc) mesh smoothing with Laplacian smoothing
+# Anders Logg 2022-11-30
+
+from fenics import *
+
+# Parameters
+n = 32
+A = 0.65
+a = 0.15
+x_0 = 0.75
+
+def smooth_simple(mesh, f):
+
+    # Create new mesh
+    new_mesh = Mesh()
+    editor = MeshEditor()
+    editor.open(new_mesh, 'triangle', 2, 2)
+    editor.init_vertices(mesh.num_vertices())
+    editor.init_cells(mesh.num_cells())
+    for i, c in enumerate(mesh.cells()):
+        editor.add_cell(i, c)
+
+    # Set new coordinates
+    for i, v in enumerate(mesh.coordinates()):
+        x, y = v
+        y += (1.0 - y)*f(x)
+        editor.add_vertex(i, (x, y))
+
+    editor.close()
+
+    return new_mesh
+
+def smooth_laplacian(mesh, f):
+
+    # Boundary value
+    class BoundaryValue(UserExpression):
+        def eval(self, values, x):
+            if near(x[1], 0.0):
+                values[0] = f(x[0])
+            else:
+                values[0] = 0.0
+        def value_shape(self):
+            return ()
+
+    # Boundary
+    boundary = 'on_boundary && (near(x[1], 0.0) || near(x[1], 1.0))'
+
+    # Solve PDE for displacement
+    parameters["reorder_dofs_serial"] = False
+    V = FunctionSpace(mesh, 'Lagrange', 1)
+    u = TrialFunction(V)
+    v = TestFunction(V)
+    a = dot(grad(u), grad(v))*dx
+    L = Constant(0)*v*dx
+    u = Function(V)
+    bv = BoundaryValue()
+    bc = DirichletBC(V, bv, boundary)
+    solve(a == L, u, bc)
+
+    # Create new mesh
+    new_mesh = Mesh()
+    editor = MeshEditor()
+    editor.open(new_mesh, 'triangle', 2, 2)
+    editor.init_vertices(mesh.num_vertices())
+    editor.init_cells(mesh.num_cells())
+    for i, c in enumerate(mesh.cells()):
+        editor.add_cell(i, c)
+
+    # Set new coordinates
+    dy = u.vector()[:]
+    for i, v in enumerate(mesh.coordinates()):
+        x, y = v
+        y += dy[i]
+        editor.add_vertex(i, (x, y))
+
+    editor.close()
+
+    return new_mesh
+
+def smooth_elastic(mesh, f):
+
+    # Boundary value for x-coordinate
+    class BoundaryValueX(UserExpression):
+        def eval(self, values, x):
+            values[0] = 0.0
+        def value_shape(self):
+            return ()
+
+    # Boundary value for y-coordinate
+    class BoundaryValueY(UserExpression):
+        def eval(self, values, x):
+            if near(x[1], 0.0):
+                values[0] = f(x[0])
+            else:
+                values[0] = 0.0
+        def value_shape(self):
+            return ()
+
+    # Boundaries
+    boundary_x = 'on_boundary && (near(x[0], 0.0) || near(x[0], 2.0) || near(x[1], 0.0))'
+    boundary_y = 'on_boundary && (near(x[1], 0.0) || near(x[1], 1.0))'
+
+    # Define elastic model
+    E = 1.0
+    nu = 0.499
+    lmbda = E*nu / ((1 + nu)*(1 - 2*nu))
[…]
