### 2022-12-01 #4928 @crvernon
📣 Mid-week rally!

👋 @pralitp and @ifthompson could you provide an update to how things are going?  Also please let me know if you have any questions.

Keep up the good work!
