### 2022-12-01 eaa76f8
add roof segmentation code
