+++ include/PointCloudProcessor.h
@@ -4,6 +4,7 @@
 #ifndef DTCC_POINT_CLOUD_PROCESSOR_H
 #define DTCC_POINT_CLOUD_PROCESSOR_H
 
+#include <Eigen/SVD>
 #include <math.h>
 #include <random>
 
@@ -16,6 +17,8 @@
 #include "PointCloud.h"
 #include "Timer.h"
 #include "Vector.h"
+
+template <typename T> int sign(T val) { return (T(0) < val) - (val < T(0)); }
 
 namespace DTCC_BUILDER
 {
@@ -355,6 +358,87 @@ public:
     return outliers;
   }
 
+  /// Returns the Distance to the K nearest neighbors for each point in
+  /// points
+  static std::vector<std::vector<double>>
+  KNNNearestNeighboursDist(std::vector<Point3D> &points, size_t neighbours)
+  {
+    size_t pc_size = points.size();
+    std::vector<std::vector<double>> neighbourDist(pc_size);
+
+    if (neighbours <= 0 or neighbours > pc_size)
+    {
+      neighbours = pc_size;
+    }
+    neighbours++; // N neighbours other than ourselves
+
+    typedef KDTreeVectorOfVectorsAdaptor<std::vector<Point3D>, double,
+                                         3 /* dims */>
+        my_kd_tree_t;
+    my_kd_tree_t pc_index(3 /*dim*/, points, 10 /* max leaf */);
+    pc_index.index->buildIndex();
+    std::vector<size_t> ret_indexes(neighbours);
+    std::vector<double> out_dists_sqr(neighbours);
+    nanoflann::KNNResultSet<double> resultSet(neighbours);
+    resultSet.init(&ret_indexes[0], &out_dists_sqr[0]);
+
+    size_t idx = 0;
+    for (auto const &pt : points)
+    {
+      std::vector<double> query_pt{pt.x, pt.y, pt.z};
+      pc_index.query(&query_pt[0], neighbours, &ret_indexes[0],
+                     &out_dists_sqr[0]);
+      for (size_t i = 1; i < neighbours;
+           i++) // start from 1 since 0 is the query point
+      {
+        neighbourDist[idx].push_back(std::sqrt(out_dists_sqr[i]));
+      }
+      idx++;
+    }
+
+    return neighbourDist;
+  }
+
+  /// Returns the index of the K nearest neighbors for each point in
+  /// points
+  static std::vector<std::vector<size_t>>
+  KNNNearestNeighboursIdx(std::vector<Point3D> &points, size_t neighbours)
+  {
+    size_t pc_size = points.size();
+    std::vector<std::vector<size_t>> neighbourIdx(pc_size);
+
+    if (neighbours <= 0 or neighbours > pc_size)
+    {
+      neighbours = pc_size;
+    }
+    neighbours++; // N neighbours other than ourselves
+
+    typedef KDTreeVectorOfVectorsAdaptor<std::vector<Point3D>, double,
+                                         3 /* dims */>
+        my_kd_tree_t;
+    my_kd_tree_t pc_index(3 /*dim*/, points, 10 /* max leaf */);
+    pc_index.index->buildIndex();
+    std::vector<size_t> ret_indexes(neighbours);
+    std::vector<double> out_dists_sqr(neighbours);
+    nanoflann::KNNResultSet<double> resultSet(neighbours);
+    resultSet.init(&ret_indexes[0], &out_dists_sqr[0]);
+    size_t idx = 0;
+    for (auto const &pt : points)
+    {
+      std::vector<double> query_pt{pt.x, pt.y, pt.z};
+      pc_index.query(&query_pt[0], neighbours, &ret_indexes[0],
+                     &out_dists_sqr[0]);
[…]
