+++ include/datamodel/Building.h
@@ -68,6 +68,9 @@ public:
   /// Array of roof points (from point cloud) inside building footprint
   std::vector<Point3D> RoofPoints{};
 
+  /// Arrary of array of
+  std::vector<std::vector<size_t>> RoofSegments{};
+
   /// Create empty building
   Building() = default;
 
@@ -104,6 +107,7 @@ public:
     error = 0;
     GroundPoints.clear();
     RoofPoints.clear();
+    RoofSegments.clear();
   }
 
   /// Pretty-print Building.

[…]
