+++ include/NetCDF4.h
@@ -1,322 +1,0 @@
-// NetCDF4 Class.
-// Copyright (C) Vasilis Naserentin
-// Licensed under the MIT License
-
-#ifndef DTCC_NETCDF4_H
-#define DTCC_NETCDF4_H
-
-#include <netcdf>
-#include <tuple>
-
-#include "Logging.h"
-
-namespace DTCC_BUILDER
-{
-using namespace netCDF;
-
-// template <class T>
-class NetCDF4
-{
-
-public:
-  std::string Name;
-  std::string FileName;
-  std::vector<double> Origin;
-  std::vector<NcVar> Coordinates;
-  // std::tuple<std::string, NcVar> CoordinatesTuple;
-  size_t Size;
-  NcVar mNcVar;
-  std::vector<size_t> CoordinateDimensions;
-  // std::unique_ptr<T[]> Pointer;
-  // T *Vector;
-  explicit NetCDF4(size_t size)
-  {
-    info("Creating new NetCDF4 with known size " + str(size));
-    // Vector = new T[size];
-    // Pointer.reset(new T[Size]);
-    // Size = size;
-  }
-  NetCDF4()
-  {
-    info("Creating new NetCDF4 without known size");
-
-    Size = 0;
-  }
-  ~NetCDF4()
-  {
-    info("Destroying NetCDF4");
-    info("Size is " + str(Size));
-    /*if (Size!=0)
-        {
-        delete Vector;
-    Vector = NULL;
-    }*/
-  }
-  template <typename T> void allocateArray(NcVar &iData)
-  {
-    std::unique_ptr<T[]> Pointer(new T[Size]);
-    T *Vector = new T[Size];
-    iData.getVar(Pointer.get());
-    iData.getVar(Vector);
-    toJson(Pointer);
-    /*if (Size != 0)
-    {
-      this->Vector = new T[Size];
-          this->Pointer.reset(new T[Size]);
-          iData.getVar(this->Vector);
-          iData.getVar(this->Pointer.get());
-    }
-    else
-    {
-      std::cout << "Size is 0. Define Size first!" << std::endl;
-    }*/
-  }
-  // void allocateArray(T *&p, size_t size, NcVar &iData) { iData.getVar(p); }
-  template <typename T> void initialize(NcVar &iVar, NcFile &iFile)
-  {
-    NcDim dim;
-    std::vector<size_t> Dimensions;
-    std::vector<std::string> CoordinateNames;
-    int numDims = iVar.getDimCount();
-    for (int i = 0; i < numDims; i++)
-    {
-      dim = iVar.getDim(i);
-      size_t dims = dim.getSize();
-      Dimensions.push_back(dims);
-
-      debug("Dim is " + str(dims) + " of " + str(numDims) + " with name " +
-            dim.getName());
-      CoordinateNames.push_back(dim.getName());
-
-      // NcVar tempvar=iFile.getVar(dim.getName());
-    }
-    Size = std::accumulate(begin(Dimensions), end(Dimensions), 1,
-                           std::multiplies<double>());
-    for (size_t i = 0; i < CoordinateNames.size(); i++)
-    {
-      NcVar tempvar = iFile.getVar(CoordinateNames[i]);
-      Coordinates.push_back(tempvar);
-      // double *tempvector = NULL;
-      // tempvector = new double[Dimensions[i]];
-      // tempvar.getVar(tempvector);
-      // CoordinatesTuple = std::make_tuple(CoordinateNames[i], tempvar);
-    }
-    CoordinateDimensions = Dimensions;
-    getOrigin(iVar);
-    allocateArray<T>(iVar);
-  }
-  void extractCoordinates(int i)
-  {
-
[…]
