+++ unittests/run-unittests.cpp
@@ -21,7 +21,5 @@ const std::string RootPath{"/home/dtcc/dtcc-builder/unittests/"};
 #include "tests/TestProtobufConversion.h"
 #include "tests/TestRoadnetworks.h"
 #include "tests/TestRoofSegmentation.h"
-#include "tests/TestSHP.h"
 #include "tests/TestUUID.h"
 #include "tests/TestUtils.h"
-#include "tests/TestXMLParser.h"

[…]
