+++ unittests/run-unittests.cpp
@@ -19,7 +19,6 @@ const std::string RootPath{"/home/dtcc/dtcc-builder/unittests/"};
 #include "tests/TestPointcloud.h"
 #include "tests/TestPolygon.h"
 #include "tests/TestProtobufConversion.h"
-#include "tests/TestRoadnetworks.h"
 #include "tests/TestRoofSegmentation.h"
 #include "tests/TestUUID.h"
 #include "tests/TestUtils.h"
