### 2022-12-05 2f3156d
add Eigen for roof segmentation
