+++ CMakeLists.txt
@@ -36,6 +36,7 @@ find_package(assimp REQUIRED)
 find_package(GDAL REQUIRED)
 find_package(PNG REQUIRED)
 find_package(LibGEOS REQUIRED)
+find_package(Eigen3 REQUIRED)
 
 # Find pybind11
 set(pybind11_DIR /usr/lib/cmake/pybind11)
@@ -82,6 +83,7 @@ include_directories(${CMAKE_SOURCE_DIR}/include/citymodel)
 include_directories(${CMAKE_SOURCE_DIR}/external)
 include_directories(${CMAKE_SOURCE_DIR}/external)
 include_directories(/usr/include/nlohmann/)
+include_directories(/usr/include/eigen3/)
 
 link_directories(/usr/local/lib) # Equivalent of ldconfig /usr/local/lib in Dockerfile
 
