### 2022-12-05 89a1990
remove more io from cpp
