+++ unittests/run-unittests.cpp
@@ -11,7 +11,6 @@ const std::string RootPath{"/home/dtcc/dtcc-builder/unittests/"};
 #include "tests/TestBoundingBox.h"
 #include "tests/TestBuilding.h"
 #include "tests/TestCityModel.h"
-#include "tests/TestColormaps.h"
 #include "tests/TestDatamodel.h"
 #include "tests/TestGEOS.h"
 #include "tests/TestGrid.h"
@@ -20,5 +19,5 @@ const std::string RootPath{"/home/dtcc/dtcc-builder/unittests/"};
 #include "tests/TestPolygon.h"
 #include "tests/TestProtobufConversion.h"
 #include "tests/TestRoofSegmentation.h"
-#include "tests/TestUUID.h"
+// #include "tests/TestUUID.h"
 #include "tests/TestUtils.h"
