### 2022-12-07 d64a93a
remove unused script
