+++ unittests/run-unittests.cpp
@@ -11,10 +11,9 @@ const std::string RootPath{"/home/dtcc/dtcc-builder/unittests/"};
 #include "tests/TestBoundingBox.h"
 #include "tests/TestBuilding.h"
 #include "tests/TestCityModel.h"
-#include "tests/TestDatamodel.h"
+// #include "tests/TestDatamodel.h"
 #include "tests/TestGEOS.h"
 #include "tests/TestGrid.h"
-#include "tests/TestParameters.h"
 #include "tests/TestPointcloud.h"
 #include "tests/TestPolygon.h"
 #include "tests/TestProtobufConversion.h"
