### 2022-12-10 a2f86ff
Added initial stuff for GS to start trying.
