+++ dtcc-sandbox/dtcc-sandbox.cpp
@@ -6,186 +6,15 @@
 
 // Needs to come before JSON (nlohmann) include because of sloppy
 // namespacing in VTK (typedef detail)...
-#include "VTK.h"
 
 // DTCC includes
-#include "CityModelGenerator.h"
-#include "CommandLine.h"
-#include "ElevationModelGenerator.h"
-#include "Geometry.h"
-#include "Grid.h"
-#include "GridField.h"
 #include "JSON.h"
-#include "LAS.h"
-#include "Logging.h"
-#include "Parameters.h"
-#include "Polygon.h"
-#include "SHP.h"
-#include "Timer.h"
-#include "VertexSmoother.h"
+#include "Mesh.h"
 
 using namespace DTCC_BUILDER;
 
 void Help() { error("Usage: dtcc-generate-citymodel Parameters.json"); }
 
-void printSet(const std::set<size_t> &setToPrint)
-{
-  for (auto It = setToPrint.cbegin(); It != setToPrint.cend(); It++)
-  {
-    std::cout << *It << " ";
-  }
-  std::cout << std::endl;
-}
-
-std::vector<std::vector<size_t>>
-getNeighbors(size_t iId,
-             const Grid2D &iGrid,
-             const std::vector<Point2D> &iCenter,
-             const std::vector<std::vector<size_t>> &iBins)
-{
-  std::vector<std::vector<size_t>> oNeighbors;
-  size_t bin = iGrid.Point2Index(iCenter[iId]);
-  std::vector<size_t> indices;
-  oNeighbors.push_back(iBins[bin]);
-  iGrid.Index2Boundary8(bin, indices);
-  // e.g. 3,4,5
-  for (size_t i = 0; i < indices.size(); i++)
-  {
-    // std::cout<<indices[i]<<std::endl;
-    oNeighbors.push_back(iBins[indices[i]]);
-  }
-  return oNeighbors;
-}
-
-std::vector<std::set<size_t>>
-getNeighborsSet(size_t iId,
-                const Grid2D &iGrid,
-                const std::vector<Point2D> &iCenter,
-                const std::vector<std::set<size_t>> &iBins)
-{
-
-  std::vector<std::set<size_t>> oNeighbors;
-  size_t bin = iGrid.Point2Index(iCenter[iId]);
-  std::vector<size_t> indices;
-  std::set<size_t> set;
-  oNeighbors.push_back(iBins[bin]);
-  iGrid.Index2Boundary8(bin, indices);
-  // e.g. 3,4,5
-  for (size_t i = 0; i < indices.size(); i++)
-  {
-    // set.insert(iBins[indices[i]]);
-
-    oNeighbors.push_back(iBins[indices[i]]);
-
-    // set.clear();
-    // std::cout<<indices[i]<<std::endl;
-    // oNeighbors.push_back(iBins[indices[i]]);
-  }
-  return oNeighbors;
-}
-
-// std::set<size_t>
-// getNeighborsSet(size_t id, const Grid2D& inGrid, const std::vector<Point2D>&
-// inCenter, const std::vector<std::set<size_t>> bins, double distance)
-// {
-//   std::set<size_t> neighbors;
-//   size_t bin = inGrid.Point2Index(inCenter[id]);
-//   std::set<size_t> set;
-//   std::vector<size_t> indices;
-//   inGrid.Index2Boundary(bin,indices);
-//   std::cout<<"Building:" << id << " is on the same bin with:"<<std::endl;
-//   for(size_t i=0;i<indices.size();i++)
-//   {
-//     std::cout<<indices[i]<<std::endl;
-//     neighbors.insert(bins[indices[i]]);
-//   }
-
-//   return neighbors;
-// }
-
-std::set<size_t> getNeighborsSet(size_t footprintId,
[…]
