### 2022-12-12 4ea52af
Computing A matrix for all cells
