+++ dtcc-sandbox/dtcc-sandbox-openmp.cpp
@@ -1,0 +1,227 @@
+// Copyright (C) 2020-2021 Anders Logg, Anton J Olsson
+// Licensed under the MIT License
+
+#include <cilk/cilk.h>
+#include <fstream>
+#include <iostream>
+#include <string>
+#include <vector>
+
+// Needs to come before JSON (nlohmann) include because of sloppy
+// namespacing in VTK (typedef detail)...
+
+// DTCC includes
+#include "JSON.h"
+#include "Mesh.h"
+#include "Timer.h"
+
+using namespace DTCC_BUILDER;
+
+void Help() { error("Usage: dtcc-sandbox CityMesh.json"); }
+
+void compute_element_matrix(double *A, const double *vertices);
+
+void compute_transformation_matrix(double *A, Mesh3D *m);
+
+int main(int argc, char *argv[])
+{
+  // Check command-line arguments
+  if (argc != 2)
+  {
+    Help();
+    return 1;
+  }
+
+  // Read parameters
+  Parameters p;
+  Mesh3D m;
+  JSON::Read(m, argv[1]);
+  info(m);
+
+  double *AA = new double[16 * m.Cells.size()];
+  std::cout << "\nComputing Element matrix for all Cells\n";
+
+  Timer t = Timer("Transformation Matrix Computation");
+  compute_transformation_matrix(AA, &m);
+  t.Stop();
+  std::cout << "\nElapsed time for A matrix Computation:" << std::endl;
+  t.Print();
+
+  // std::ofstream outfile ("test.txt");
+  // for (size_t i = 0; i < 16 * m.Cells.size(); i++)
+  // {
+  //   outfile << AA[i] << std::endl;
+  // }
+  // outfile.close();
+
+  std::cout << "\nNumber of Cells: " << m.Cells.size() << std::endl;
+  delete[] AA;
+  return 0;
+}
+
+// Output: A = array of length 16
+// Represents flattened 4x4 matrix
+//
+// Input: vertices = array of length 12
+// Represents flattened 4x3 matrix of vertex coordinates
+
+void compute_element_matrix(double *A, const double *vertices)
+{
+  const double J_c4 = vertices[7] - vertices[1];
+  const double J_c8 = vertices[11] - vertices[2];
+  const double J_c5 = vertices[10] - vertices[1];
+  const double J_c7 = vertices[8] - vertices[2];
+  const double J_c0 = vertices[3] - vertices[0];
+  const double J_c1 = vertices[6] - vertices[0];
+  const double J_c6 = vertices[5] - vertices[2];
+  const double J_c3 = vertices[4] - vertices[1];
+  const double J_c2 = vertices[9] - vertices[0];
+
+  alignas(32) double sp[80];
+  sp[0] = J_c4 * J_c8;
+  sp[1] = J_c5 * J_c7;
+  sp[2] = sp[0] + -1 * sp[1];
+  sp[3] = J_c0 * sp[2];
+  sp[4] = J_c5 * J_c6;
+  sp[5] = J_c3 * J_c8;
+  sp[6] = sp[4] + -1 * sp[5];
+  sp[7] = J_c1 * sp[6];
+  sp[8] = sp[3] + sp[7];
+  sp[9] = J_c3 * J_c7;
[…]
