+++ dtcc-generate-roadnetwork/roadnetwork_pb2.py
@@ -1,0 +1,266 @@
+# Generated by the protocol buffer compiler.  DO NOT EDIT!
+# source: roadnetwork.proto
+
+import sys
+_b=sys.version_info[0]<3 and (lambda x:x) or (lambda x:x.encode('latin1'))
+from google.protobuf import descriptor as _descriptor
+from google.protobuf import message as _message
+from google.protobuf import reflection as _reflection
+from google.protobuf import symbol_database as _symbol_database
+# @@protoc_insertion_point(imports)
+
+_sym_db = _symbol_database.Default()
+
+
+
+
+DESCRIPTOR = _descriptor.FileDescriptor(
+  name='roadnetwork.proto',
+  package='RoadNetwork',
+  syntax='proto3',
+  serialized_options=None,
+  serialized_pb=_b('\n\x11roadnetwork.proto\x12\x0bRoadNetwork\" \n\x08Vector2D\x12\t\n\x01x\x18\x01 \x01(\x02\x12\t\n\x01y\x18\x02 \x01(\x02\"A\n\x0cGeoreference\x12\x0b\n\x03\x63rs\x18\x01 \x01(\t\x12\x0c\n\x04\x65psg\x18\x02 \x01(\x05\x12\n\n\x02x0\x18\x03 \x01(\x02\x12\n\n\x02y0\x18\x04 \x01(\x02\"^\n\x04Road\x12\x10\n\x08vertices\x18\x01 \x03(\x05\x12\x0c\n\x04type\x18\x02 \x01(\x05\x12\x0c\n\x04name\x18\x03 \x01(\t\x12\r\n\x05width\x18\x04 \x01(\x02\x12\r\n\x05lanes\x18\x05 \x01(\x05\x12\n\n\x02id\x18\x06 \x01(\t\"\x87\x01\n\x0bRoadNetwork\x12%\n\x06vertex\x18\x01 \x03(\x0b\x32\x15.RoadNetwork.Vector2D\x12 \n\x05roads\x18\x02 \x03(\x0b\x32\x11.RoadNetwork.Road\x12/\n\x0cgeoreference\x18\x03 \x01(\x0b\x32\x19.RoadNetwork.Georeferenceb\x06proto3')
+)
+
+
+
+
+_VECTOR2D = _descriptor.Descriptor(
+  name='Vector2D',
+  full_name='RoadNetwork.Vector2D',
+  filename=None,
+  file=DESCRIPTOR,
+  containing_type=None,
+  fields=[
+    _descriptor.FieldDescriptor(
+      name='x', full_name='RoadNetwork.Vector2D.x', index=0,
+      number=1, type=2, cpp_type=6, label=1,
+      has_default_value=False, default_value=float(0),
+      message_type=None, enum_type=None, containing_type=None,
+      is_extension=False, extension_scope=None,
+      serialized_options=None, file=DESCRIPTOR),
+    _descriptor.FieldDescriptor(
+      name='y', full_name='RoadNetwork.Vector2D.y', index=1,
+      number=2, type=2, cpp_type=6, label=1,
+      has_default_value=False, default_value=float(0),
+      message_type=None, enum_type=None, containing_type=None,
+      is_extension=False, extension_scope=None,
+      serialized_options=None, file=DESCRIPTOR),
+  ],
+  extensions=[
+  ],
+  nested_types=[],
+  enum_types=[
+  ],
+  serialized_options=None,
+  is_extendable=False,
+  syntax='proto3',
+  extension_ranges=[],
+  oneofs=[
+  ],
+  serialized_start=34,
+  serialized_end=66,
+)
+
+
+_GEOREFERENCE = _descriptor.Descriptor(
+  name='Georeference',
[…]

[…]
