### 2022-12-14 a42b962
WIP: cmake hacking
