### 2022-12-29 ce8cefb
WIP remove io from C++
