### 2023-01-05 #4928 @ifthompson
I've worked my way through the checklist. Documentation is very nice and I was able to install and run successfully. Only thing I did not find was clear Community Guidelines for how users can contribute, report issues, and get support (I assume this is just raise an issue on the repository, so include a sentence saying that, or point to broader guidelines if they are elsewhere in the DTCC project).

A few other notes:

It might be helpful to include that tip for Windows users about the Unix-style line endings in the Installation section, in addition (or instead) of the Development section. Also, there may be a .gitattributes setting that tells Git not to convert line endings.

There is a typo in paper.md, "algorihtms" is misspelled in the Method and implementation first paragraph last sentence  https://gitlab.com/dtcc-platform/dtcc-builder/-/blob/joss-2022/doc/paper-joss-2022/paper.md#method-and-implementation
