### 2023-01-09 f933105
Remove documentation about fixing line endings since this is now handled by .gitattributes
