### 2023-01-09 ab11284
Add notes on how to contribute as suggested by JOSS reviewer
