+++ unittests/run-unittests.cpp
@@ -8,15 +8,15 @@
 const std::string RootPath{"/home/dtcc/dtcc-builder/unittests/"};
 
 // Run these tests. You can comment out any test you don't want run
-// #include "tests/TestBoundingBox.h"
-// #include "tests/TestBuilding.h"
-// #include "tests/TestCityModel.h"
+#include "tests/TestBoundingBox.h"
+#include "tests/TestBuilding.h"
+#include "tests/TestCityModel.h"
 // #include "tests/TestDatamodel.h"
-// #include "tests/TestGEOS.h"
-// #include "tests/TestGrid.h"
-// #include "tests/TestPointcloud.h"
-// #include "tests/TestPolygon.h"
+#include "tests/TestGEOS.h"
+#include "tests/TestGrid.h"
+#include "tests/TestPointcloud.h"
+#include "tests/TestPolygon.h"
 #include "tests/TestProtobufConversion.h"
-// #include "tests/TestRoofSegmentation.h"
+#include "tests/TestRoofSegmentation.h"
 // #include "tests/TestUUID.h"
-// #include "tests/TestUtils.h"
+#include "tests/TestUtils.h"
