+++ dtcc-pybuilder/dtcc-pybindings/src/pybind_builder.cpp
@@ -17,7 +17,7 @@
 #include "Logging.h"
 #include "Mesh.h"
 #include "MeshGenerator.h"
-#include "MeshIO.h"
+// #include "MeshIO.h"
 #include "MeshProcessor.h"
 #include "Point.h"
 #include "PointCloud.h"
@@ -288,15 +288,6 @@ Surface3D MergeSurfaces3D(const std::vector<Surface3D> &surfaces)
   return merged_surface;
 }
 
-bool WriteSurface3D(const Surface3D &surface,
-                    std::string fileName,
-                    std::string format,
-                    bool YUp)
-{
-  MeshIO::Write(surface, fileName, format, YUp);
-  return true;
-}
-
 bool WriteVTKMesh3D(const Mesh3D &mesh, std::string filepath)
 {
   VTK::Write(mesh, filepath);
@@ -501,9 +492,6 @@ PYBIND11_MODULE(_pybuilder, m)
 
   m.def("MergeSurfaces3D", &DTCC_BUILDER::MergeSurfaces3D,
         "Merge surfaces into a single surface");
-
-  m.def("WriteSurface3D", &DTCC_BUILDER::WriteSurface3D,
-        "Write surface to file");
 
   m.def("WriteVTKMesh3D", &DTCC_BUILDER::WriteVTKMesh3D,
         "Write 3D mesh to VTK format");
