### 2023-02-24 66447b0
fix install script
