+++ docker/install_scripts/install_dependencies.sh
@@ -37,10 +37,22 @@ apt-get update && apt-get install -y \
     python3-pip \
     python3-dev \
     libprotobuf-dev \
+    libassimp-dev \
     protobuf-compiler
+
+#these seem necessary for the python bindings to work
+#these seem necessary for the python bindings to work
+if [-f /usr/lib/x86_64-linux-gnu/libassimp.so.5]; then
+    ln -s /usr/lib/x86_64-linux-gnu/libassimp.so.5 /usr/lib/
+fi
+if [-f /usr/lib/aarch64-linux-gnu/libassimp.so.5]; then
+    ln -s /usr/lib/aarch64-linux-gnu/libassimp.so.5 /usr/lib/
+fi
+
 
 # ./install_assimp.sh
 ./install_libgeos.sh
 # ./install_libLAS.sh
 ./install_VTK.sh
 ./install_py_libs.sh
+./install_dtccio.sh
