+++ dtcc-pybuilder/dtcc/builder/PointCloud.py
@@ -13,14 +13,14 @@ from typing import List, Tuple
 class PointCloud:
     """Class for storing point cloud object"""
 
-    def __init__(self, pointcloud_file=None, pb_sting = None, bounds=()):
+    def __init__(self, pointcloud_path=None, pb_sting = None, bounds=()):
         """ """
         self._builder_pc = None
         self.origin = (0, 0)
         self.bounds = bounds
         self.used_classifications = {}
-        if pointcloud_file is not None:
-            self.load_from_path(pointcloud_file)
+        if pointcloud_path is not None:
+            self.load_from_path(pointcloud_path)
         if pb_sting is not None:
             self.from_protobuf(pb_sting)
 
