+++ docker/install_scripts/install_py_libs.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/bash
 
 python3 -m pip install \
-    numpy \
+    numpy>=1.23.* \
     laspy[lazrs] \
     protobuf==3.20.* \
 	pybind11==2.10.* \
