+++ docker/install_scripts/install_dependencies.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+#add-apt-repository -y ppa:ubuntugis/ubuntugis-unstable
+#apt-get update
 
 apt-get update && apt-get install -y \
     build-essential \
@@ -41,17 +43,18 @@ apt-get update && apt-get install -y \
     protobuf-compiler
 
 #these seem necessary for the python bindings to work
-#these seem necessary for the python bindings to work
-if [-f /usr/lib/x86_64-linux-gnu/libassimp.so.5]; then
+if [ -f /usr/lib/x86_64-linux-gnu/libassimp.so.5 ]; then
     ln -s /usr/lib/x86_64-linux-gnu/libassimp.so.5 /usr/lib/
 fi
-if [-f /usr/lib/aarch64-linux-gnu/libassimp.so.5]; then
+if [ -f /usr/lib/aarch64-linux-gnu/libassimp.so.5]; then
     ln -s /usr/lib/aarch64-linux-gnu/libassimp.so.5 /usr/lib/
 fi
 
+# remove old version of libgeos and install new version from source
+apt remove -y libgeos-dev
+./install_libgeos.sh
 
 # ./install_assimp.sh
-./install_libgeos.sh
 # ./install_libLAS.sh
 ./install_VTK.sh
 ./install_py_libs.sh
