+++ docker/install_scripts/install_dtccmodel.sh
@@ -1,0 +1,11 @@
+git clone https://github.com/dtcc-platform/dtcc-model.git
+cd dtcc-model
+git checkout dev/new-build-system
+
+# make sure we rebuild the protobuf files using the
+# same verion of protoc as installed in the docker image
+rm -rf src/cpp/protobuf/dtcc.pb.*
+mkdir build
+cd build
+cmake ..
+make && make install
