+++ src/dtcc_builder/ElevationModel.py
@@ -1,6 +1,6 @@
 from google.protobuf.json_format import MessageToJson
 
-from dtcc.builder import _pybuilder
+from dtcc_builder import _pybuilder
 
 
 class ElevationModel:
