+++ tests/test_citymodel.py
@@ -7,13 +7,13 @@ sys.path.append(str((Path(__file__).parent / "..").resolve()))
 import json
 
 from dtcc import builder
-from dtcc import io
+import dtcc_io as io
 
-from dtcc.builder import _pybuilder
+from dtcc_builder import _pybuilder
 
-from dtcc.builder import CityModel
-from dtcc.builder import ElevationModel
-from dtcc.builder import PointCloud
+from dtcc_builder import CityModel
+from dtcc_builder import ElevationModel
+from dtcc_builder import PointCloud
 
 data_dir = (Path(__file__).parent / "../../unittests/data").resolve()
 p = builder.Parameters.load_parameters(data_dir / "MinimalCase" / "Parameters.json")
