+++ src/dtcc_builder/__init__.py
@@ -1,8 +1,8 @@
 
-from dtcc.builder.CityModel import CityModel
-from dtcc.builder.ElevationModel import ElevationModel
-from dtcc.builder.PointCloud import PointCloud
-from dtcc.builder.PointCloud import PointCloud
-from dtcc.builder import _pybuilder
+from dtcc_builder.CityModel import CityModel
+from dtcc_builder.ElevationModel import ElevationModel
+from dtcc_builder.PointCloud import PointCloud
+from dtcc_builder.PointCloud import PointCloud
+from dtcc_builder import _pybuilder
 
-from dtcc.builder.Parameters import load_parameters
+from dtcc_builder.Parameters import load_parameters
