+++ src/cpp/include/Laplacian.h
@@ -1,5 +1,5 @@
 // This code conforms with the UFC specification version 2018.1.0
-// and was automatically generated by FFC version 2019.1.0.post0.
+// and was automatically generated by FFC version 2019.2.0.dev0.
 //
 // This code was generated with the option '-l dolfin' and
 // contains DOLFIN-specific wrappers that depend on DOLFIN.
@@ -120,7 +120,7 @@ public:
                                 const double * X) const final override
   {
     static const double coefficients0[1][4] = { { 0.288675134594813, -0.18257418583505536, -0.10540925533894598, -0.07453559924999298 } };
-    static const double coefficients1[1][4] = { { 0.28867513459481287, 0.18257418583505536, -0.10540925533894596, -0.07453559924999296 } };
+    static const double coefficients1[1][4] = { { 0.2886751345948129, 0.18257418583505536, -0.10540925533894596, -0.07453559924999296 } };
     static const double coefficients2[1][4] = { { 0.2886751345948129, 0.0, 0.21081851067789195, -0.07453559924999299 } };
     static const double coefficients3[1][4] = { { 0.2886751345948129, 0.0, 0.0, 0.22360679774997896 } };
     for (std::size_t k = 0; k < num_points * 4; ++k)
@@ -167,21 +167,21 @@ public:
     if (order > 1)
         return;
     // Tables of derivatives of the polynomial base (transpose).
-    alignas(32) static const double dmats0[3][4][4] = {
-        {{0.0, 0.0, 0.0, 0.0},
-         {6.324555320336758, 0.0, 0.0, 0.0},
-         {0.0, 0.0, 0.0, 0.0},
-         {0.0, 0.0, 0.0, 0.0}},
-        {{0.0, 0.0, 0.0, 0.0},
-         {3.162277660168379, 0.0, 0.0, 0.0},
-         {5.477225575051663, 0.0, 0.0, 0.0},
-         {0.0, 0.0, 0.0, 0.0}},
-        {{0.0, 0.0, 0.0, 0.0},
-         {3.162277660168379, 0.0, 0.0, 0.0},
-         {1.8257418583505545, 0.0, 0.0, 0.0},
-         {5.163977794943224, 0.0, 0.0, 0.0}}};
+    alignas(32) static const double dmats0[3][4][4] =
+        { { { 0.0, 0.0, 0.0, 0.0 },
+            { 6.324555320336757, 0.0, 0.0, 0.0 },
+            { 0.0, 0.0, 0.0, 0.0 },
+            { 0.0, 0.0, 0.0, 0.0 } },
+          { { 0.0, 0.0, 0.0, 0.0 },
+            { 3.1622776601683786, 0.0, 0.0, 0.0 },
+            { 5.477225575051662, 0.0, 0.0, 0.0 },
+            { 0.0, 0.0, 0.0, 0.0 } },
+          { { 0.0, 0.0, 0.0, 0.0 },
+            { 3.1622776601683786, 0.0, 0.0, 0.0 },
+            { 1.825741858350554, 0.0, 0.0, 0.0 },
+            { 5.163977794943222, 0.0, 0.0, 0.0 } } };
     static const double coefficients0[1][4] = { { 0.288675134594813, -0.18257418583505536, -0.10540925533894598, -0.07453559924999298 } };
-    static const double coefficients1[1][4] = { { 0.28867513459481287, 0.18257418583505536, -0.10540925533894596, -0.07453559924999296 } };
+    static const double coefficients1[1][4] = { { 0.2886751345948129, 0.18257418583505536, -0.10540925533894596, -0.07453559924999296 } };
     static const double coefficients2[1][4] = { { 0.2886751345948129, 0.0, 0.21081851067789195, -0.07453559924999299 } };
[…]
