+++ docker/install_scripts/install_VTK.sh
@@ -4,7 +4,7 @@ tar xvzf VTK-7.1.1.tar.gz
 cd VTK-7.1.1/ 
 mkdir build
 cd build
-cmake -DVTK_Group_Rendering=OFF -DVTK_BUILD_ALL_MODULES_FOR_TESTS:BOOL=OFF -DVTK_Group_StandAlone=OFF -DModule_vtkCommonCore:BOOL=ON -DModule_vtkCommonDataModel:BOOL=ON -DModule_vtkIOXML:BOOL=ON .. 
+cmake -DVTK_Group_Rendering=OFF -DVTK_BUILD_ALL_MODULES_FOR_TESTS:BOOL=OFF -DVTK_Group_StandAlone=OFF -DModule_vtkCommonCore:BOOL=ON -DModule_vtkCommonDataModel:BOOL=ON -DVTK_Group_Qt:BOOL=OFF -DModule_vtkIOXML:BOOL=ON .. 
 make all -j 4
 sudo make install
 
