+++ docker/install_scripts/install_dependencies.sh
@@ -9,6 +9,7 @@ DEBIAN_FRONTEND=noninteractive apt-get update && apt-get install -y \
     libtriangle-dev \
     nlohmann-json3-dev \
     libpthread-stubs0-dev \
+    libuuid1 \
     libeigen3-dev \
     libgdal-dev \
     python3 \
@@ -31,13 +32,13 @@ DEBIAN_FRONTEND=noninteractive apt-get update && apt-get install -y \
 
 # apt-get install -y python3-fiona python3-rasterio
 
-# #these seem necessary for the python bindings to work
-# if [ -f /usr/lib/x86_64-linux-gnu/libassimp.so.5 ]; then
-#     ln -s /usr/lib/x86_64-linux-gnu/libassimp.so.5 /usr/lib/
-# fi
-# if [ -f /usr/lib/aarch64-linux-gnu/libassimp.so.5]; then
-#     ln -s /usr/lib/aarch64-linux-gnu/libassimp.so.5 /usr/lib/
-# fi
+#these seem necessary for the python bindings to work
+if [ -f /usr/lib/x86_64-linux-gnu/libassimp.so.5 ]; then
+    ln -s /usr/lib/x86_64-linux-gnu/libassimp.so.5 /usr/lib/
+fi
+if [ -f /usr/lib/aarch64-linux-gnu/libassimp.so.5]; then
+    ln -s /usr/lib/aarch64-linux-gnu/libassimp.so.5 /usr/lib/
+fi
 
 # remove old version of libgeos and install new version from source
 # apt remove -y libgeos-dev
