### 2023-03-08 afa2711
Fix polygon simplification
