+++ scripts/dtcc-convert-shp2geojson/CMakeLists.txt
@@ -1,4 +1,0 @@
-set(PROGRAM dtcc-convert-shp2geojson)
-add_executable(${PROGRAM} ${PROGRAM}.cpp)
-target_link_libraries(${PROGRAM} pugixml proj shp las)
-install(TARGETS ${PROGRAM} DESTINATION bin)
