+++ scripts/dtcc-convert-geojson2shp/CMakeLists.txt
@@ -1,4 +1,0 @@
-set(PROGRAM dtcc-convert-geojson2shp)
-add_executable(${PROGRAM} ${PROGRAM}.cpp)
-target_link_libraries(${PROGRAM} pugixml proj shp las)
-install(TARGETS ${PROGRAM} DESTINATION bin)
