+++ pyproject.toml
@@ -22,7 +22,8 @@ classifiers = [
 dependencies = [
  "pybind11 >= 2.10.0, < 2.11.0",
  "numpy",
- "dtcc-io@git+https://github.com/dtcc-platform/dtcc-io.git@dev/new-build-system"
+ "dtcc-model@git+https://github.com/dtcc-platform/dtcc-model.git@develop",
+ "dtcc-io@git+https://github.com/dtcc-platform/dtcc-io.git@develop"
 ]
 
 [project.optional-dependencies]
