+++ dtcc-generate-mesh/dtcc-generate-mesh.cpp
@@ -1,294 +1,0 @@
-// Copyright (C) 2020 Anders Logg
-// Licensed under the MIT License
-
-#include <dolfin.h>
-
-#include "CityModelGenerator.h"
-#include "CommandLine.h"
-#include "GridField.h"
-#include "JSON.h"
-#include "LaplacianSmoother.h"
-#include "Logging.h"
-#include "Mesh.h"
-#include "MeshGenerator.h"
-#include "MeshIO.h"
-#include "MeshProcessor.h"
-#include "OBJ.h"
-#include "ParameterProcessor.h"
-#include "Parameters.h"
-#include "Utils.h"
-#include "VTK.h"
-#include "VertexSmoother.h"
-
-#include "LAS.h"
-#include "SHP.h"
-
-using namespace DTCC_BUILDER;
-
-void Help() { error("Usage: dtcc-generate-mesh Parameters.json"); }
-
-// Generate surface meshes (non-matching, used for visualization)
-void GenerateSurfaceMeshes(const CityModel &cityModel,
-                           const GridField2D &dtm,
-                           const Parameters &p)
-{
-  // Get data directory
-  const std::string dataDirectory = p["DataDirectory"];
-
-  // Get origin (for serialization purposes)
-  Point2D origin({p["X0"], p["Y0"]});
-
-  // Generate mesh for ground and buildings
-  Surface3D groundSurface;
-  std::vector<Surface3D> buildingSurfaces;
-  MeshGenerator::GenerateSurfaces3D(groundSurface, buildingSurfaces, cityModel,
-                                    dtm, p["MeshResolution"]);
-
-  // Merge building surfaces
-  Surface3D buildingSurface;
-  MeshProcessor::MergeSurfaces3D(buildingSurface, buildingSurfaces);
-
-  // Write to file
-  if (p["WriteJSON"])
-  {
-    JSON::Write(groundSurface, dataDirectory + "GroundSurface.json", origin);
-    JSON::Write(buildingSurface, dataDirectory + "BuildingSurface.json",
-                origin);
-  }
-
-  // Write data for debugging and visualization
-  if (p["WriteVTK"])
-  {
-    VTK::Write(groundSurface, dataDirectory + "GroundSurface.vtu");
-    VTK::Write(buildingSurface, dataDirectory + "BuildingSurface.vtu");
-  }
-
-  if (p["WriteSTL"])
-  {
-    MeshIO::Write(groundSurface, dataDirectory + "GroundSurface.stl", "stl",
-                  false);
-    MeshIO::Write(buildingSurface, dataDirectory + "BuildingSurface.stl", "stl",
-                  false);
-  }
-
-  if (p["WriteOBJ"])
-  {
-    MeshIO::Write(groundSurface, dataDirectory + "GroundSurface.obj", "obj",
-                  true);
-    MeshIO::Write(buildingSurface, dataDirectory + "BuildingSurface.obj", "obj",
-                  true);
-  }
-}
-
-// Generate volume meshes (matching, used for simulation)
-void GenerateVolumeMeshes(CityModel &cityModel,
-                          const GridField2D &dtm,
-                          const Parameters &p)
-{
-  // Get directories
-  const std::string dataDirectory = p["DataDirectory"];
-  const std::string outputDirectory = p["OutputDirectory"];
-
-  // Get origin (for serialization purposes)
-  Point2D origin({p["X0"], p["Y0"]});
-
-  // Step 1: Generate city model (and elevation model).
[…]
