+++ dtcc-generate-citymodel/dtcc-generate-citymodel.cpp
@@ -1,211 +1,0 @@
-// Copyright (C) 2020-2022 Anders Logg, Anton J Olsson, Dag Wästberg
-// Licensed under the MIT License
-
-#include <experimental/filesystem>
-#include <string>
-#include <vector>
-
-// Needs to come before JSON (nlohmann) include because of sloppy
-// namespacing in VTK (typedef detail)...
-#include "VTK.h"
-
-// DTCC includes
-#include "CityModelGenerator.h"
-#include "CommandLine.h"
-#include "ElevationModelGenerator.h"
-#include "GridField.h"
-#include "JSON.h"
-#include "LAS.h"
-#include "Logging.h"
-#include "ParameterProcessor.h"
-#include "Parameters.h"
-#include "Polygon.h"
-#include "SHP.h"
-#include "Timer.h"
-#include "Utils.h"
-#include "VertexSmoother.h"
-
-using namespace DTCC_BUILDER;
-
-void Help() { error("Usage: dtcc-generate-citymodel Parameters.json"); }
-
-int main(int argc, char *argv[])
-{
-  progress(0.0);
-
-  // Check command-line arguments
-  if (CommandLine::HasOption("-h", argc, argv))
-  {
-    Help();
-    return 0;
-  }
-  Parameters p = ParameterProcessor::ProcessArgs(argc, argv);
-
-  // Get directories
-  const std::string dataDirectory = p["DataDirectory"];
-  const std::string outputDirectory = p["OutputDirectory"];
-  info("Loding data from directory: " + dataDirectory);
-  info("Saving data to directory:   " + outputDirectory);
-
-  // Start timer
-  Timer timer("Step 1: Generate city model");
-
-  // Read property map
-  std::vector<Polygon> footprints;
-  std::vector<std::string> UUIDs;
-  std::vector<int> entityIDs;
-
-  auto loading_timer = Timer("load data");
-  SHP::Read(footprints, dataDirectory + "PropertyMap.shp", &UUIDs, &entityIDs);
-  info("Loaded " + str(footprints.size()) + " building footprints");
-
-  // Set bounding box
-  BoundingBox2D bbox;
-  Point2D O;
-  if (p["AutoDomain"])
-  {
-    bbox = BoundingBox2D(footprints, p["DomainMargin"]);
-    info("Bounding box of footprints: " + str(bbox));
-    BoundingBox2D lasBBox;
-    LAS::BoundsDirectory(lasBBox, dataDirectory);
-    info("Bounding box of point cloud: " + str(lasBBox));
-    bbox.Intersect(lasBBox);
-    O = bbox.P;
-    p["X0"] = O.x;
-    p["Y0"] = O.y;
-    p["XMin"] = 0.0;
-    p["YMin"] = 0.0;
-    p["XMax"] = bbox.Q.x - bbox.P.x;
-    p["YMax"] = bbox.Q.y - bbox.P.y;
-  }
-  else
-  {
-    O = Point2D(p["X0"], p["Y0"]);
-    const double xMin = p["XMin"];
-    const double xMax = p["XMax"];
-    const double yMin = p["YMin"];
-    const double yMax = p["YMax"];
-    const Point2D P{O.x + xMin, O.y + yMin};
-    const Point2D Q{O.x + xMax, O.y + yMax};
-    bbox = BoundingBox2D(P, Q);
-  }
-
-  // Check size of bounding box
-  info("Bounding box of city model: " + str(bbox));
-  if (bbox.Area() < 100.0)
-  {
-    error("Domain too small to generate a city model");
-    return 1;
-  }
-
-  // Read point cloud (only points inside bounding box)
-  auto loadlas_timer = Timer("Reaing LAS");
-  PointCloud pointCloud;
[…]
