+++ dtcc-generate-citymodel/CMakeLists.txt
@@ -1,4 +1,0 @@
-set(PROGRAM dtcc-generate-citymodel)
-add_executable(${PROGRAM} ${PROGRAM}.cpp)
-target_link_libraries(${PROGRAM} ${VTK_LIBRARIES} pugixml proj shp las uuid stdc++fs)
-install(TARGETS ${PROGRAM} DESTINATION bin)
