+++ pyproject.toml
@@ -29,6 +29,10 @@ dependencies = [
 [project.optional-dependencies]
 test = ["pytest"]
 
+[project.scripts]
+dtcc-build = "dtcc_builder.scripts:dtcc_build.main"
+
+
 [tool.scikit-build]
 wheel.expand-macos-universal-tags = true
 
