+++ src/dtcc_builder/Meshing.py
@@ -59,16 +59,18 @@ def extract_open_surface3D(boundary: _pybuilder.Surface3D) -> _pybuilder.Surface
 def merge_surfaces3D(surfaces: List[_pybuilder.Surface3D]) -> _pybuilder.Surface3D:
     return _pybuilder.MergeSurfaces3D(surfaces)
 
+
 def write_Protobuf_surface3D(surface: _pybuilder.Surface3D, out_file):
     with open(out_file, "wb") as f:
         pb = _pybuilder.convertSurface3DToProtobuf(surface)
         f.write(pb)
 
 
-def write_surface(surface:_pybuilder.Surface3D, file_name):
+def write_surface(surface: _pybuilder.Surface3D, file_name):
     pb = _pybuilder.convertSurface3DToProtobuf(surface)
     io.mesh.write(file_name, pb)
 
-def write_volume_mesh(mesh:_pybuilder.Mesh3D, file_name):
+
+def write_volume_mesh(mesh: _pybuilder.Mesh3D, file_name):
     pb = _pybuilder.convertMesh3DToProtobuf(mesh)
     io.mesh.write(file_name, pb, volume_mesh=True)
