+++ README.md
@@ -11,10 +11,10 @@ supported by Sweden’s Innovation Agency Vinnova under Grant No. 2019-421 00041
 
 ## Documentation
 
-* [Introduction](./doc/introduction.md)
-* [Installation](./doc/installation.md)
-* [Usage](./doc/usage.md)
-* [Development](./doc/development.md)
+* [Introduction](./docs/introduction.md)
+* [Installation](./docs/installation.md)
+* [Usage](./docs/usage.md)
+* [Development](./docs/development.md)
 
 ## Authors (in order of appearance)
 
