+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "scikit_build_core.build"
 [project]
 name = "dtcc-builder"
 version = "0.5.1"
-description="DTCC CityModel Builder"
+description="DTCC Builder"
 readme = "README.md"
 authors = [
   { name = "Anders Logg", email = "logg@chalmers.se" },
@@ -34,4 +34,3 @@ dtcc-build = "dtcc_builder.scripts:dtcc_build.main"
 
 [tool.scikit-build]
 wheel.expand-macos-universal-tags = true
-
