+++ dtcc-sandbox/CMakeLists.txt
@@ -1,6 +1,0 @@
-set(PROGRAM dtcc-sandbox)
-include(${VTK_USE_FILE})
-link_directories(/usr/local/lib)
-add_executable(${PROGRAM} ${PROGRAM}.cpp)
-target_link_libraries(${PROGRAM} ${VTK_LIBRARIES} pugixml proj shp las uuid stdc++fs)
-install(TARGETS ${PROGRAM} DESTINATION bin)
