+++ bench/README.md
@@ -1,0 +1,1 @@
+These benchmarks are currently broken due to redesign waiting to be fixed.
