+++ src/cpp/include/MeshIO.h
@@ -1,149 +1,0 @@
-// Copyright (C) 2020-2021 Dag Wästberg
-// Licensed under the MIT License
-
-#ifndef DTCC_MESHIO_H
-#define DTCC_MESHIO_H
-
-// Assimp includes
-#include <assimp/Exporter.hpp>
-#include <assimp/Importer.hpp>
-#include <assimp/material.h>
-#include <assimp/mesh.h>
-#include <assimp/scene.h>
-#include <assimp/vector3.h>
-
-#include "Logging.h"
-#include "Mesh.h"
-#include "Point.h"
-#include "Simplex.h"
-#include "Surface.h"
-
-namespace DTCC_BUILDER
-{
-
-class MeshIO
-{
-public:
-  static void Write(const Surface3D &surface,
-                    std::string fileName,
-                    std::string format,
-                    bool YUp = false)
-  {
-    info("Assimp writing 3D surface to file " + fileName + " in format " +
-         format);
-
-    Assimp::Exporter *surfaceExporter = new Assimp::Exporter();
-    aiScene *scene = new aiScene();
-    if (scene)
-    {
-      aiNode *rootNode = new aiNode();
-      rootNode->mNumMeshes = 1;
-      rootNode->mMeshes = new uint[1];
-      rootNode->mMeshes[0] = 0;
-      scene->mRootNode = rootNode;
-
-      scene->mMaterials = new aiMaterial *[1];
-      scene->mMaterials[0] = defaultMaterial();
-      scene->mNumMaterials = 1;
-
-      scene->mFlags = 8;
-      scene->mNumMeshes = 1;
-      scene->mMeshes = new aiMesh *[1];
-      scene->mMeshes[0] = buildAIMesh(surface.Vertices, surface.Faces, YUp);
-
-      int result = surfaceExporter->Export(scene, format, fileName);
-      if (result == 0)
-      {
-        info("File exported successfully: " + fileName);
-      }
-      else
-      {
-        error("There was an error in file export");
-        error("Error:" + str(surfaceExporter->GetErrorString()));
-      }
-    }
-    else
-    {
-      error("Assimp Error in creating scene");
-    }
-  }
-
-private:
-  static aiMaterial *defaultMaterial()
-  {
-    aiMaterial *material = new aiMaterial();
-
-    material->AddProperty(new aiVector3D(1, 1, 1), 1, AI_MATKEY_COLOR_DIFFUSE);
-    return material;
-  }
-
-  static aiMesh *buildAIMesh(std::vector<Point3D> vertices,
-                             std::vector<Simplex2D> faces,
-                             bool YUp)
-  {
-    aiMesh *mesh = new aiMesh();
-    mesh->mPrimitiveTypes = aiPrimitiveType_TRIANGLE;
-    mesh->mNumVertices = vertices.size();
-    mesh->mNumFaces = faces.size();
-
-    aiVector3D *aiVertices = new aiVector3D[mesh->mNumVertices];
-    for (size_t i = 0; i < mesh->mNumVertices; i++)
-    {
-      auto p = vertices[i];
-      if (YUp)
-        aiVertices[i] = aiVector3D(p.x, p.z, p.y);
-      else
-        aiVertices[i] = aiVector3D(p.x, p.y, p.z);
-    }
-    mesh->mVertices = aiVertices;
-
-    aiFace *aiFaces = new aiFace[mesh->mNumFaces];
-
-    aiVector3D V1, V2, V3;
-    double poly_sum;
-    for (size_t i = 0; i < mesh->mNumFaces; i++)
-    {
[…]
