+++ src/cpp/include/Timer.h
@@ -9,6 +9,7 @@
 #include <utility>
 
 // #include "JSON.h"
+#include "Table.h"
 
 namespace DTCC_BUILDER
 {
