+++ src/dtcc_builder/Build.py
@@ -1,0 +1,192 @@
+import dtcc_builder as builder
+from dtcc_builder import _pybuilder
+import dtcc_io as io
+from pathlib import Path
+
+
+def project_domain(p):
+    if p["AutoDomain"]:
+        footprint_bounds = io.citymodel.building_bounds(
+            p["DataDirectory"] / p["BuildingsFileName"], p["DomainMargin"]
+        )
+        las_bounds = io.pointcloud.calc_las_bounds(p["PointCloudDirectory"])
+        domain_bounds = io.bounds.bounds_intersect(footprint_bounds, las_bounds)
+        origin = domain_bounds[:2]
+        p["X0"] = origin[0]
+        p["Y0"] = origin[1]
+        p["XMin"] = 0.0
+        p["YMin"] = 0.0
+        p["XMax"] = domain_bounds[2] - domain_bounds[0]
+        p["YMax"] = domain_bounds[3] - domain_bounds[1]
+    else:
+        origin = (p["X0"], p["Y0"])
+        domain_bounds = (
+            p["X0"] + p["XMin"],
+            p["Y0"] + p["YMin"],
+            p["X0"] + p["XMax"],
+            p["Y0"] + p["YMax"],
+        )
+    return (origin, domain_bounds)
+
+
+def create_citymodel(
+    building_footprint_path, pointcloud_path, parameters=None, return_protobuf=True
+):
+    """Create a city model from a directory of point clouds and a shapefile of building footprints.
+    Args:
+    @param building_footprint_path: Path to a shapefile containing building footprints
+    @param pointcloud_path: Path to a pointcloud file or a directory containing point clouds
+    @param parameters: parameters dict or Path to a JSON file containing parameters for the city model.
+    If None default parameters will be used.
+    @param return_protobuf: If True, the city model will be returned as a protobuf string
+
+    @return: A tuple containing the city model and elevation model as a protobuf string
+    (if return_protobuf is True) or as builder objects (if return_protobuf is False)
+    """
+    if type(parameters) is dict:
+        p = parameters
+    else:
+        p = builder.load_parameters(parameters)
+
+    origin, domain_bounds = project_domain(p)
+    if io.bounds.bounds_area(domain_bounds) < 100:
+        raise ValueError("Domain too small to generate a city model")
+    pc = builder.PointCloud(pointcloud_path=pointcloud_path, bounds=domain_bounds)
+    if len(pc) == 0:
+        raise ValueError("No points in point cloud")
+    pc.set_origin(origin)
+    pc.remove_global_outliers(p["OutlierMargin"])
+    if p["NaiveVegitationFilter"]:
+        pc.vegetation_filter()
+    cm = builder.CityModel(
+        footprints_file=building_footprint_path, parameters=p, bounds=domain_bounds
+    )
+    cm.set_origin(origin)
+    cm.clean_citymodel()
+
+    dtm = builder.ElevationModel(pc, p["ElevationModelResolution"], [2, 9])
+    # dsm = ElevationModel(pc, p["ElevationModelResolution"])
+    dtm.smooth_elevation_model(p["GroundSmoothing"])
+
+    cm.extract_building_points(pc)
+
+    if 6 not in pc.used_classifications and p["RANSACOutlierRemover"]:
+        cm.building_points_RANSAC_outlier_remover()
+
+    if p["StatisticalOutlierRemover"]:
+        cm.building_points_statistical_outlier_remover()
+
+    cm.compute_building_heights(dtm)
+
+    if return_protobuf:
+        cm = cm.to_protobuf()
+        dtm = dtm.to_protobuf()
+    return (cm, dtm)
+
+
[…]
