+++ src/dtcc_builder/scripts/dtcc_build.py
@@ -11,15 +11,10 @@ from pathlib import Path
 import os
 import re
 
+from google.protobuf.json_format import MessageToJson
+
 import dtcc_io as io
 import dtcc_builder as builder
-from dtcc_builder import CityModel
-from dtcc_builder import ElevationModel
-from dtcc_builder import PointCloud
-from dtcc_builder import _pybuilder
-from dtcc_builder import load_parameters
-from dtcc_builder import Meshing
-
 
 def camel_to_kebab(name):
     name = re.sub("(.)([A-Z][a-z]+)", r"\1-\2", name)
@@ -30,7 +25,7 @@ def create_parameters_options(parser):
     type_map = {"str": str, "float": float, "int": int, "bool": bool}
 
     name_translator = {}
-    p = load_parameters()
+    p = builder.load_parameters()
     for k, v in p.items():
         kebab_name = camel_to_kebab(k)
         name_translator[k] = kebab_name
@@ -76,33 +71,6 @@ def get_project_paths(path):
 
     return (parameters_file, project_path)
 
-
-def calc_project_domain(p):
-    if p["AutoDomain"]:
-        footprint_bounds = io.citymodel.building_bounds(
-            p["DataDirectory"] / p["BuildingsFileName"], p["DomainMargin"]
-        )
-        las_bounds = io.pointcloud.calc_las_bounds(p["PointCloudDirectory"])
-        domain_bounds = io.bounds.bounds_intersect(
-            footprint_bounds, las_bounds)
-        origin = domain_bounds[:2]
-        p["X0"] = origin[0]
-        p["Y0"] = origin[1]
-        p["XMin"] = 0.0
-        p["YMin"] = 0.0
-        p["XMax"] = domain_bounds[2] - domain_bounds[0]
-        p["YMax"] = domain_bounds[3] - domain_bounds[1]
-    else:
-        origin = (p["X0"], p["Y0"])
-        domain_bounds = (
-            p["X0"] + p["XMin"],
-            p["Y0"] + p["YMin"],
-            p["X0"] + p["XMax"],
-            p["Y0"] + p["YMax"],
-        )
-    return (origin, domain_bounds)
-
-
 def set_directories(p, project_path):
     if p["DataDirectory"] == "":
         p["DataDirectory"] = project_path
@@ -122,116 +90,7 @@ def set_directories(p, project_path):
 
     return p
 
-
-def generate_citymodel(p):
-    origin, domain_bounds = calc_project_domain(p)
-
-    if io.bounds.bounds_area(domain_bounds) < 100:
-        print("Domain too small to generate a city model")
-        sys.exit(1)
-
-    pc = builder.PointCloud(
-        pointcloud_path=p["PointCloudDirectory"], bounds=domain_bounds
-    )
-    if len(pc) == 0:
-        print("Error: Point cloud in domain is empty")
-        sys.exit(1)
-    pc.set_origin(origin)
-
-    pc.remove_global_outliers(p["OutlierMargin"])
-
-    if p["NaiveVegitationFilter"]:
-        pc.vegetation_filter()
-
-    cm = CityModel(p["DataDirectory"] / p["BuildingsFileName"])
-    cm.set_origin(origin)
-    cm.clean_citymodel()
-
-    dtm = ElevationModel(pc, p["ElevationModelResolution"], [2, 9])
-    # dsm = ElevationModel(pc, p["ElevationModelResolution"])
-    dtm.smooth_elevation_model(p["GroundSmoothing"])
-
-    cm.extract_building_points(pc)
-
-    # 6 is building classification
[…]
