+++ src/dtcc_builder/__init__.py
@@ -6,3 +6,5 @@ from dtcc_builder.PointCloud import PointCloud
 from dtcc_builder import _pybuilder
 
 from dtcc_builder.Parameters import load_parameters
+
+from dtcc_builder.Build import create_citymodel, create_surface_meshes, create_mesh 
