+++ src/dtcc_builder/CityModel.py
@@ -173,7 +173,10 @@ class CityModel:
 
     def to_protobuf(self) -> str:
         """convert CityModel to protobuf string"""
-        return _pybuilder.convertCityModelToProtobuf(self._builder_cm)
+        pb_string = _pybuilder.convertCityModelToProtobuf(self._builder_cm)
+        
+        return pb_string
+    
 
 
 def get_building_footprint(building: _pybuilder.Building):
