### 2023-03-17 057cbb7
fix protobuf loading
