+++ src/dtcc_builder/__init__.py
@@ -3,6 +3,7 @@ from dtcc_builder.CityModel import CityModel
 from dtcc_builder.ElevationModel import ElevationModel
 from dtcc_builder.PointCloud import PointCloud
 from dtcc_builder.PointCloud import PointCloud
+from dtcc_builder import Meshing
 from dtcc_builder import _pybuilder
 
 from dtcc_builder.Parameters import load_parameters
