### 2023-03-17 95a2449
simplify loading parameters
