+++ src/dtcc_builder/__init__.py
@@ -8,4 +8,4 @@ from dtcc_builder import _pybuilder
 
 from dtcc_builder.Parameters import load_parameters
 
-from dtcc_builder.Build import create_citymodel, create_surface_meshes, create_mesh 
+from dtcc_builder.Build import build_citymodel, build_surface_meshes, build_mesh
