+++ src/dtcc_builder/Meshing.py
@@ -68,9 +68,9 @@ def write_Protobuf_surface3D(surface: _pybuilder.Surface3D, out_file):
 
 def write_surface(surface: _pybuilder.Surface3D, file_name):
     pb = _pybuilder.convertSurface3DToProtobuf(surface)
-    io.mesh.write(file_name, pb)
+    io.write_mesh(file_name, pb)
 
 
 def write_volume_mesh(mesh: _pybuilder.Mesh3D, file_name):
     pb = _pybuilder.convertMesh3DToProtobuf(mesh)
-    io.mesh.write(file_name, pb, volume_mesh=True)
+    io.write_mesh(file_name, pb)
