### 2023-03-17 d364992
simplify write_mesh
