+++ tests/python/test_meshing.py
@@ -125,7 +125,7 @@ class TestMeshing(unittest.TestCase):
         boundary = Meshing.extract_boundary3D(mesh3D)
 
         out_file = "test.obj"
-        done = Meshing.write_surface(boundary, out_file, "obj")
+        done = Meshing.write_mesh(boundary, out_file)
         self.assertTrue(done)
         self.assertTrue(os.path.isfile(out_file))
         os.unlink(out_file)
