### 2023-03-21 9fc4d03
Update ci.yml
Trying to always cleanup!
