### 2023-03-22 0415292
make test data path relative
