### 2023-03-22 0c45afa
Remove old remnant of doc folder
