+++ doc/paper-joss-2022/paper.bib
@@ -1,85 +1,0 @@
-
-@book{geoscontributorsGEOSCoordinateTransformation2021,
-  title = {{{GEOS}} Coordinate Transformation Software Library},
-  author = {{GEOS contributors}},
-  year = {2021},
-  publisher = {{Open Source Geospatial Foundation}}
-}
-
-@article{ketzlerDigitalTwinsCities2020,
-  title = {Digital {{Twins}} for {{Cities}}: {{A State}} of the {{Art Review}}},
-  author = {Ketzler, Bernd and Naserentin, Vasilis and Latino, Fabio and Zangelidis, Christopher and Thuvander, Liane and Logg, Anders},
-  year = {2020},
-  journal = {Built Environment},
-  volume = {46},
-  number = {4},
-  pages = {547--573},
-  issn = {0263-7960},
-  doi = {10.2148/benv.46.4.547},
-  abstract = {During the last decades, a variety of digital tools have been developed to support both the planning and management of cities, as well as the inclusion of civic society. Here, the concept of a Digital Twin \textendash{} which is rapidly emerging throughout many disciplines due to advances in technology, computational capacities and availability of large amounts of data \textendash{} plays an important role. In short, a digital twin is a living virtual model, a connected digital representation of a physical system and has been a central concept in the manufacturing industry for the past decades. In this article, we review the terminology of digital twins for cities and identify commonalities and relations to the more established term 3D city models. Our fi ndings indicate an increasing use of the term digital twin in academic literature, both in general and in the context of cities and the built environment. We fi nd that while there is as yet no consensus on the exact defi nition of what constitutes a digital twin, it is increasingly being used to describe something that is more than a 3D city model (including, e.g. semantic data, real-time sensor data, physical models, and simulations). At the same time, the term has not yet replaced the term 3D city model as the most dominant term in the 3D GIS domain. By looking at grey literature we discuss how digital twins for cities are implemented in practice and present examples of digital twins in a global perspective. Further, we discuss some of the application areas and potential challenges for future development and implementation of digital twins for cities. We conclude that there are signifi cant opportunities for up-scaling digital twins, with the potential to bring benefi ts to the city and its citizens and clients.}
-}
-
-@article{ledoux3dfierAutomaticReconstruction2021,
-  title = {3dfier: Automatic Reconstruction of {{3D}} City Models},
-  author = {Ledoux, Hugo and Biljecki, Filip and Dukai, Bal{\'a}zs and Kumar, Kavisha and Peters, Ravi and Stoter, Jantien and Commandeur, Tom},
-  year = {2021},
-  month = jan,
-  journal = {Journal of Open Source Software},
-  volume = {6},
-  number = {57},
-  pages = {2866},
-  publisher = {{The Open Journal}},
-  issn = {2475-9066},
-  doi = {10.21105/joss.02866}
-}
-
-@article{ledouxCityJSONCompactEasytouse2019,
-  title = {{{CityJSON}}: A Compact and Easy-to-Use Encoding of the {{CityGML}} Data Model},
-  author = {Ledoux, Hugo and Arroyo Ohori, Ken and Kumar, Kavisha and Dukai, Bal{\'a}zs and Labetski, Anna and Vitalis, Stelios},
-  year = {2019},
-  journal = {Open Geospatial Data, Software and Standards},
-  issn = {2363-7501},
-  doi = {10.1186/s40965-019-0064-0},
[…]
