### 2023-03-24 #4928 @ipadjen
Hello! I checked the repository and I have a few questions both for the authors and the editor.

@crvernon one checklist point states:
'Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.'
My question: Is the list of dependencies necessary if the program is meant to be run through a docker container?

I initially tried using the default branch ('main') but that one did not work -- the docker container would not build. The branch 'joss-2022' was the one that worked well. I installed the repository using documentation and ran the example -- everything went fine. Afterwards, I tested the code with two of my own datasets (not in Sweden) and everything seems to be running fine -- the documentation is clear, it is straightforward to use the code and I have not had robustness issues.

With that in mind, I would like to address a few key points that I found relevant.

First, a few general notes:
- I see the information on GitLab that the project has been moved to GitHub. Should the repository information be updated? GitLab and the link to the archived repository are on line 48 of the paper.
- .obj surface export seems to be broken. It produces meshes that are rotated 90 degrees over the x-axis with flipped normals.
- I have not managed to find information on point cloud classification. Should it be classified (if yes, what classes are mandatory?) or can the code work its way around unclassified point clouds? It would be nice to include that information in the [input data](https://gitlab.com/dtcc-platform/dtcc-builder/-/blob/joss-2022/doc/usage.md#input-data).
- a note on one of the author's names in the references of the paper. It happens to be me, and there is 'd' (Paden) missing in there :)

Finally, here are my questions and remarks about the two main functionalities:
- The first one is to generate meshes (surfaces) suitable for visualisation. The results mostly look nice; however, it seems that holes in polygons are ignored (see figure below). Is this an oversight?
- The second is to create computational grids for simulations. Do you happen to have an example of a simulation/results using the input model from the DTCC builder? I think it is important to show that the code is capable of creating grids for simulations, such as grids adhering to best practice guidelines (https://www.sciencedirect.com/science/article/pii/S0360132315000724), or a grid convergence study. I could not test this feature myself and I got the impression it is supposed to be a strong (or even the strongest) contribution by the project.

![poly_with_holes](https://user-images.githubusercontent.com/49401914/227593874-8ca0c903-3676-46a9-aa85-7094926c3066.png)
