### 2023-03-28 #4928 @anderslogg
@ipadjen Here's some initial response to some of your comments (more to follow).

Regarding the repo, yes we have moved to GitHub since we submitted the paper. So please use the GitHub repo for posting issues. The branch name is still `joss-2022`. Many things are different in the `develop` branch, in particular the addition of a Python interface (in progress).

Regarding the spelling of your name, we have this in the `.bib` file: `Pa{\dj}en` Is that not correct? We tried hard to get it right. :-)

Comments that remain to be handled:

1. `.obj` export broken
2. Requirements for point cloud classification
3. Holes not handled
4. Examples of using the meshes for simulation
