### 2023-03-28 dtcc-platform/dtcc-builder#11 @ipadjen
OBJ export broken
Hello! I'm opening this issue as a part of the JOSS review: https://github.com/openjournals/joss-reviews/issues/4928

There is a problem with OBJ export. It seems that there are two issues:

1.  The results are rotated 90 degrees over the x-axis.
2.  Normals are flipped for all faces.

Here are the results of the example case HelsingborgResidential and CitySurface in .stl, .vtu, and .obj:
