### 2023-03-28 dtcc-platform/dtcc-builder#12 @ipadjen
Handling polygons with holes
Hello,

An issue from the JOSS review: https://github.com/openjournals/joss-reviews/issues/4928

Polygons with holes seem to be ignored:

![poly_with_holes](https://user-images.githubusercontent.com/49401914/227593874-8ca0c903-3676-46a9-aa85-7094926c3066.png)
