### 2023-03-28 dtcc-platform/dtcc-builder#13 @ipadjen
Add requirements for point clouds to documentation
Hello, as a part of JOSS review: https://github.com/openjournals/joss-reviews/issues/4928

It would be nice to add few sentences on point cloud requirements to the documentation, more specifically, the information on classification (which classes are necessary, if any).
