### 2023-03-28 dtcc-platform/dtcc-builder#14 @ipadjen
An example of a simulation with an input from DTCC-builder
Hello, as a part of JOSS review: https://github.com/openjournals/joss-reviews/issues/4928

It would be nice to show in the paper that grids can be used to run simulations. I think that a grid sensitivity analysis would be good example.

Let me know if you have any other ideas.
