### 2023-03-28 #4928 @ipadjen
Great! There, I opened up issues.

Yeah, about the name, I keep having this issue with a non-standard character. It is correct, but you also need an additional package in latex to make it work. I started writing 'd' instead myself, it's just easier.
