+++ src/dtcc_builder/scripts/dtcc_build.py
@@ -16,6 +16,7 @@ from google.protobuf.json_format import MessageToJson
 import dtcc_io as io
 import dtcc_builder as builder
 
+
 def camel_to_kebab(name):
     name = re.sub("(.)([A-Z][a-z]+)", r"\1-\2", name)
     return re.sub("([a-z0-9])([A-Z])", r"\1-\2", name).lower()
@@ -32,8 +33,7 @@ def create_parameters_options(parser):
         name_translator[kebab_name] = k
         val_type = type(v).__name__
         if val_type == "bool":
-            parser.add_argument(f"--{kebab_name}",
-                                default=v, action="store_true")
+            parser.add_argument(f"--{kebab_name}", default=v, action="store_true")
         else:
             parser.add_argument(
                 f"--{kebab_name}", default=None, type=type_map[val_type]
@@ -43,7 +43,6 @@ def create_parameters_options(parser):
 
 def update_parameters_from_options(p, args, name_translator):
     for k, v in p.items():
-        print(k)
         snake_name = name_translator.get(k)
         if snake_name is None:
             continue
@@ -71,8 +70,8 @@ def get_project_paths(path):
 
     return (parameters_file, project_path)
 
-def run(p, citymodel_only, mesh_only):
 
+def run(p, citymodel_only):
     print(p)
     #cm, dtm = None, None
     footprint_bounds = io.citymodel.building_bounds(
@@ -82,36 +81,40 @@ def run(p, citymodel_only, mesh_only):
     pointcloud = io.load_pointcloud(p["PointCloudDirectory"], footprint_bounds)
     city_model = builder.build_citymodel(footprints, pointcloud, p)
     if p["WriteJSON"]:
-        with open(p["OutputDirectory"]/ "CityModel.json", "w") as dst:
+        with open(output_path / "CityModel.json", "w") as dst:
             dst.write(MessageToJson(city_model))
     if p["WriteProtobuf"]:
-        with open(p["OutputDirectory"]/ "CityModel.pb", "wb") as dst:
-            dst.write(city_model.SerializeToString())
-    io.save_citymodel(city_model, p["OutputDirectory"] / "CityModel.shp",)
+        with open(output_path / "CityModel.pb", "wb") as dst:
+            dst.write(cm.SerializeToString())
+    io.save_citymodel(
+        cm,
+        output_path / "CityModel.shp",
+    )
     if not citymodel_only:
-        volume_mesh, surface_mesh = builder.build_mesh(city_model,p )
+        volume_mesh, surface_mesh = builder.build_mesh(cm, dtm, p)
 
         if p["WriteJSON"]:
-            with open(p["OutputDirectory"]/ "CitySurface.json", "w"):
+            with open(output_path / "CitySurface.json", "w"):
                 MessageToJson(surface_mesh)
-            with open(p["OutputDirectory"]/ "CityMesh.json", "w"):
+            with open(output_path / "CityMesh.json", "w"):
                 MessageToJson(volume_mesh)
-        
+
         if p["WriteProtobuf"]:
-            with open(p["OutputDirectory"]/ "CitySurface.pb", "wb") as dst:
+            with open(output_path / "CitySurface.pb", "wb") as dst:
                 dst.write(surface_mesh.SerilaizeToString())
-            with open(p["OutputDirectory"]/ "CityMesh.pb", "wb") as dst:
+            with open(output_path / "CityMesh.pb", "wb") as dst:
                 dst.write(volume_mesh.SerializeToString())
         if p["WriteVTK"]:
-            io.write_mesh(surface_mesh,p["OutputDirectory"]/ "CitySurface.vtk")
-            io.write_mesh(volume_mesh,p["OutputDirectory"]/ "CityMesh.vtk")
[…]
