+++ doc/paper-joss-2022/paper.bib
@@ -64,7 +64,7 @@
 
 @article{padenAutomaticReconstruction3D2022,
   title = {Towards Automatic Reconstruction of {{3D}} City Models Tailored for Urban Flow Simulations},
-  author = {Pa{\dj}en, Ivan and {Garc{\'i}a-S{\'a}nchez}, Clara and Ledoux, Hugo},
+  author = {Paden, Ivan and {Garc{\'i}a-S{\'a}nchez}, Clara and Ledoux, Hugo},
   year = {2022},
   journal = {Frontiers in Built Environment},
   volume = {8},
@@ -81,5 +81,3 @@
   pages = {203--222},
   publisher = {{Springer}}
 }
-
-
