### 2023-03-29 df36658
Add point note on point cloud classification requirements.
