+++ doc/usage.md
@@ -27,10 +27,10 @@ The output data may be found in the corresponding subdirectory of the
 and how to visualize the generated city models and meshes are
 described in detail below.
 
-> **Note:** Running `dtcc-generate-mesh` on an Apple M1 processor is 
-> currently not fully supported due to problems with the M1 processor and 
-> the triangle library. 
- 
+> **Note:** Running `dtcc-generate-mesh` on an Apple M1 processor is
+> currently not fully supported due to problems with the M1 processor and
+> the triangle library.
+
 > **Note:** To run the demo, you first need to build and install DTCC Builder. You
 must also download the demo datasets as described in the
 [installation instructions](./installation.md) for DTCC Builder.
@@ -69,6 +69,12 @@ If a parameter file is given as argument, the specified
 `DataDirectory` parameter is searched for the input data:
 
     dtcc-generate-citymodel <path to parameter file>
+
+> **Note:** The minimal requirement for point cloud data is that
+ground points are classified as 2 with other points that are not
+ground having a different classification (not 2). If classifications
+for buildings (6), vegetation (3-5) and/or water (9) are present,
+those are used to potentially improve the city model generation.
 
 ### Output data
 
@@ -228,5 +234,3 @@ by the `GroundSmoothing` parameter:
 experimental parameters may be added/removed during development. For
 a complete list of  parameters, refer to the parameter files
 `dtcc-generate-[citymodel,mesh].json` generated by running the demo.
-
-
