### 2023-03-29 dtcc-platform/dtcc-builder#13 @anderslogg
Added the following to the docs (usage.md):

> **Note:** The minimal requirement for point cloud data is that
ground points are classified as 2 with other points that are not
ground having a different classification (not 2). If classifications
for buildings (6), vegetation (3-5) and/or water (9) are present,
those are used to potentially improve the city model generation.
