+++ sandbox/smoothing-2023/JSON.h
@@ -1,0 +1,1355 @@
+// Copyright (C) 2020-2021 Anders Logg, Anton J Olsson
+// Licensed under the MIT License
+
+#ifndef DTCC_JSON_H
+#define DTCC_JSON_H
+
+#include <datamodel/CityModel.h>
+//#include <datamodel/District.h>
+#include <fstream>
+#include <iostream>
+#include <nlohmann/json.hpp>
+#include <utility>
+
+#include "BoundingBox.h"
+#include "Color.h"
+//#include "ColorMap.h"
+#include "Grid.h"
+#include "GridField.h"
+#include "GridVectorField.h"
+#include "Mesh.h"
+//#include "Parameters.h"
+#include "PointCloud.h"
+#include "Surface.h"
+//#include "cityjson/CityJSON.h"
+#include "datamodel/RoadNetwork.h"
+
+namespace DTCC
+{
+
+  class JSON
+  {
+  public:
+
+    // Utility functions for JSON input/output
+    #include "JSONUtils.h"
+
+    /// Deserialize Parameters
+    static void Deserialize(Parameters& parameters, const nlohmann::json& json)
+    {
+      CheckType("Parameters", json);
+
+      // Read parameter values
+      for (auto &it : parameters.Map)
+      {
+        if (!HasKey(it.first, json))
+          warning("Missing parameter \"" + it.first +
+                  "\", using default value");
+        else
+        {
+          switch (it.second.Type)
+          {
+          case ParameterType::Bool:
+            it.second = static_cast<bool>(json[it.first]);
+            break;
+          case ParameterType::Int:
+            it.second = static_cast<int>(json[it.first]);
+            break;
+          case ParameterType::Float:
+            it.second = static_cast<double>(json[it.first]);
+            break;
+          case ParameterType::String:
+            std::string valueString = json[it.first];
+            it.second = valueString;
+            break;
+          }
+        }
+      }
+
+      // Report unknown parameters
+      for (auto it = json.begin(); it != json.end(); ++it)
+      {
+        if (!parameters.HasKey(it.key()) && it.key() != "Type")
+          warning("Unknown parameter \"" + it.key() + "\", ignoring");
+      }
+    };
+
+    /// Serialize Parameters
+    static void Serialize(const Parameters& parameters, nlohmann::json& json)
+    {
+      json["Type"] = "Parameters";
+      for (auto &it : parameters.Map)
+      {
+        switch (it.second.Type)
+        {
+        case ParameterType::Bool:
+          json[it.first] = static_cast<bool>(it.second);
+          break;
+        case ParameterType::Int:
+          json[it.first] = static_cast<int>(it.second);
+          break;
+        case ParameterType::Float:
+          json[it.first] = static_cast<double>(it.second);
+          break;
+        case ParameterType::String:
+          json[it.first] = static_cast<std::string>(it.second);
+          break;
+        }
+      }
+    }
+
+    /// Serialize timings (special-case, only output supported)
+    static void
+    Serialize(std::map<std::string, std::pair<double, size_t>> timings,
+              nlohmann::json &json)
+    {
+      for (const auto &it : timings)
+      {
+        const std::string task = it.first;
+        const double total = it.second.first;
[…]
