### 2023-04-04 #4928 @ipadjen
@crvernon I am still wondering whether the list of all dependencies is necessary or not. The checklist item explicitly states its necessity, although I see little benefit to it with this repo. It's all in there in the docker anyway. What do you think?
