### 2023-04-04 dtcc-platform/dtcc-builder#11 @ipadjen
I gave it a look, this is what I see:
- Face orientations are fixed.
- CitySurface is still 90 degrees rotated, whereas the other .objs are fine. I noticed it might be a design decision due to Blender?
- The results still look different from .stls and vtus. Could it be that vertex normals are still weirdly oriented? OBJ is on the right of the figure.

![obj_comparison](https://user-images.githubusercontent.com/49401914/229870625-85e17f25-70e2-42b9-9c33-caf9a8294ab8.png)
