### 2023-04-05 dtcc-platform/dtcc-builder#11 @vbassn
@ipadjen Thanks for this! Checking again with my coauthor @dwastberg
