### 2023-04-05 dtcc-platform/dtcc-builder#11 @vbassn
@ipadjen Which viewer are you using for the last screenshot? Is it Paraview?
