### 2023-04-05 dtcc-platform/dtcc-builder#11 @ipadjen
Yes, the screenshot is from Paraview. I also checked it in MeshLab and a similar thing -- STL vs OBJ are colored differently.
