+++ include/Parameters.h
@@ -31,7 +31,7 @@ public:
     Add("WriteJSON", true);
     Add("WriteVTK", true);
     Add("WriteOBJ", true);
-    Add("WriteSTL", true);
+    // Add("WriteSTL", true);
     Add("WriteMatrix", false);
     Add("Debug", false);
 
