+++ sandbox/smoothing-2023/include/boundaryConditions.hpp
@@ -1,0 +1,194 @@
+
+#ifndef BOUNDARY_CONDITIONS_HPP
+#define BOUNDARY_CONDITIONS_HPP
+
+#include <algorithm>
+#include <fstream>
+#include <iostream>
+#include <iterator>
+#include <string>
+#include <vector>
+
+#include "GridField.h"
+#include "LaplacianSmoother.h"
+#include "Mesh.h"
+#include "datamodel/CityModel.h"
+
+void getVerticeMarkers(const Mesh3D &mesh, int *vMarkers)
+{
+  const size_t nC = mesh.Cells.size();
+  const size_t nV = mesh.Vertices.size();
+
+  for (size_t i = 0; i < nV; i++)
+  {
+    vMarkers[i] = -4;
+  }
+
+  std::array<uint, 4> I = {0};
+
+  size_t k0 = 0;
+  size_t k1 = 0;
+  size_t k2 = 0;
+  size_t k3 = 0;
+  for (size_t cn = 0; cn < nC; cn++)
+  {
+    // Initializing Global Index for each cell
+    I[0] = mesh.Cells[cn].v0;
+    I[1] = mesh.Cells[cn].v1;
+    I[2] = mesh.Cells[cn].v2;
+    I[3] = mesh.Cells[cn].v3;
+
+    const double z_mean = (mesh.Vertices[I[0]].z + mesh.Vertices[I[1]].z +
+                           mesh.Vertices[I[2]].z + mesh.Vertices[I[3]].z) /
+                          4;
+
+    const int cellMarker = mesh.Markers[cn];
+
+    if (cellMarker >= 0) // Building
+    {
+      for (size_t i = 0; i < 4; i++)
+      {
+        if (mesh.Vertices[I[i]].z > z_mean)
+        {
+          continue;
+        }
+        vMarkers[I[i]] = cellMarker;
+      }
+    }
+    else if (cellMarker == -1) // Building Halo
+    {
+      for (size_t i = 0; i < 4; i++)
+      {
+        if (mesh.Vertices[I[i]].z > z_mean)
+        {
+          continue;
+        }
+        vMarkers[I[i]] = std::max(vMarkers[I[i]], -1);
+      }
+    }
+    else if (cellMarker == -2) // Ground
+    {
+      for (size_t i = 0; i < 4; i++)
+      {
+        if (mesh.Vertices[I[i]].z > z_mean)
+        {
+          continue;
+        }
+        vMarkers[I[i]] = std::max(vMarkers[I[i]], -2);
+      }
+    }
+    else if (cellMarker == -3) // Top
+    {
+      for (size_t i = 0; i < 4; i++)
+      {
+        if (mesh.Vertices[I[i]].z < z_mean)
+        {
+          continue;
+        }
+        vMarkers[I[i]] = std::max(vMarkers[I[i]], -3);
+      }
+    }
+  }
+  for (size_t v = 0; v < nV; v++)
+  {
+    if (vMarkers[v] >= 0)
+      k0++;
+    else if (vMarkers[v] == -1)
+      k1++;
+    else if (vMarkers[v] == -2)
+      k2++;
+    else if (vMarkers[v] == -3)
+      k3++;
+  }
+
+  int k4 = nV - (k0 + k1 + k2 + k3);
+  std::cout << "k0 :" << k0 << std::endl
+            << "k1 :" << k1 << std::endl
+            << "k2 :" << k2 << std::endl
+            << "k3 :" << k3 << std::endl
+            << "k4 :" << k4 << std::endl
+            << k0 + k1 + k2 + k3 + k4 << "  " << nV << std::endl;
+
+  return;
+}
+
+void applyBC_load(const int *vMarkers,
+                  const Mesh3D &mesh,
[…]
