+++ .github/workflows/ci.yml
@@ -36,3 +36,10 @@ jobs:
     # This fails after checkout/v3 was not used
     #- name: more cleanup ops
     #  run : pwd && sudo rm -rf /home/ubuntu/Runner/actions-runner/_work/${{ github.event.repository.name }}
+    - name : dispatch
+      if: success()
+      run: |
+            curl -L -X POST -H "Accept: application/vnd.github+json" \
+            -H "Authorization: Bearer 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"\
+            -H "X-GitHub-Api-Version: 2022-11-28" https://api.github.com/repos/dtcc-platform/dtcc-docs/actions/workflows/manual.yml/dispatches \
+            -d '{"ref":$BRANCH,"inputs":{"name": "Me"}}'
