+++ .github/workflows/ci.yml
@@ -30,12 +30,6 @@ jobs:
     #  run: cd ${{ github.event.repository.name }}/build/ && cmake .. && make all && sudo make install && cd ..
     - name: run tests
       run: cd ${{ github.event.repository.name }}/tests && pwd && ls -alt && sh run-tests
-    - name: cleanup op
-      if: always()
-      run : sudo find /home/ubuntu/Runner/actions-runner/_work/${{ github.event.repository.name }}/. -name . -o -prune -exec rm -rf -- {} + || true
-    # This fails after checkout/v3 was not used
-    #- name: more cleanup ops
-    #  run : pwd && sudo rm -rf /home/ubuntu/Runner/actions-runner/_work/${{ github.event.repository.name }}
     - name : dispatch
       if: always()
       run: |
@@ -43,3 +37,10 @@ jobs:
             -H "Authorization: Bearer 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"\
             -H "X-GitHub-Api-Version: 2022-11-28" https://api.github.com/repos/dtcc-platform/dtcc-docs/actions/workflows/manual.yml/dispatches \
             -d '{"ref":$BRANCH,"inputs":{"name": "Me"}}'
+    - name: cleanup op
+      if: always()
+      run : sudo find /home/ubuntu/Runner/actions-runner/_work/${{ github.event.repository.name }}/. -name . -o -prune -exec rm -rf -- {} + || true
+    # This fails after checkout/v3 was not used
+    #- name: more cleanup ops
+    #  run : pwd && sudo rm -rf /home/ubuntu/Runner/actions-runner/_work/${{ github.event.repository.name }}
+    
