+++ .github/workflows/ci.yml
@@ -34,7 +34,7 @@ jobs:
       if: always()
       run: |
             curl -L -X POST -H "Accept: application/vnd.github+json" \
-            -H "Authorization: Bearer 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"\
+            -H "Authorization: Bearer $TOKEN"\
             -H "X-GitHub-Api-Version: 2022-11-28" https://api.github.com/repos/dtcc-platform/dtcc-docs/actions/workflows/manual.yml/dispatches \
             -d '{"ref":$BRANCH,"inputs":{"name": "Me"}}'
     - name: cleanup op
