+++ .github/workflows/ci.yml
@@ -8,6 +8,7 @@ on:
 
 env:
   BRANCH: develop
+  ACCESS_TOKEN: ${{ secrets.TOKEN }}
 
 jobs:
   build:
@@ -34,7 +35,7 @@ jobs:
       if: always()
       run: |
             curl -L -X POST -H "Accept: application/vnd.github+json" \
-            -H "Authorization: Bearer $TOKEN"\
+            -H "Authorization: Bearer $ACCESS_TOKEN"\
             -H "X-GitHub-Api-Version: 2022-11-28" https://api.github.com/repos/dtcc-platform/dtcc-docs/actions/workflows/manual.yml/dispatches \
             -d '{"ref":$BRANCH,"inputs":{"name": "Me"}}'
     - name: cleanup op
