+++ .github/workflows/ci.yml
@@ -8,7 +8,6 @@ on:
 
 env:
   BRANCH: develop
-  ACCESS_TOKEN: ${{ secrets.TOKEN }}
 
 jobs:
   build:
@@ -33,11 +32,13 @@ jobs:
       run: cd ${{ github.event.repository.name }}/tests && pwd && ls -alt && sh run-tests
     - name : dispatch
       if: always()
-      run: |
-            curl -L -X POST -H "Accept: application/vnd.github+json" \
-            -H "Authorization: Bearer $ACCESS_TOKEN"\
-            -H "X-GitHub-Api-Version: 2022-11-28" https://api.github.com/repos/dtcc-platform/dtcc-docs/actions/workflows/manual.yml/dispatches \
-            -d '{"ref":$BRANCH,"inputs":{"name": "Me"}}'
+      run: curl \
+           -X POST \
+           -H "Accept:application/vnd.github+json" \
+           -H "Authorization:token ${{ secrets.TOKEN }}" \
+           -H "X-GitHub-Api-Version:2022-11-28" \
+           https://api.github.com/repos/dtcc-platform/dtcc-docs/actions/workflows/manual.yml/dispatches \
+           -d '{"ref":$BRANCH,"inputs":{"name":"Me"}}'
     - name: cleanup op
       if: always()
       run : sudo find /home/ubuntu/Runner/actions-runner/_work/${{ github.event.repository.name }}/. -name . -o -prune -exec rm -rf -- {} + || true
