+++ .github/workflows/ci.yml
@@ -38,7 +38,7 @@ jobs:
            -H "Authorization:token ${{ secrets.TOKEN }}" \
            -H "X-GitHub-Api-Version:2022-11-28" \
            https://api.github.com/repos/dtcc-platform/dtcc-docs/actions/workflows/manual.yml/dispatches \
-           -d '{"ref":$BRANCH,"inputs":{"name":"Me"}}'
+           -d '{"ref":develop,"inputs":{"name":"Me"}}'
     - name: cleanup op
       if: always()
       run : sudo find /home/ubuntu/Runner/actions-runner/_work/${{ github.event.repository.name }}/. -name . -o -prune -exec rm -rf -- {} + || true
