+++ docs/development.md
@@ -63,3 +63,17 @@ not convert Unix-style file endings on checkout. This can be
 accomplished by:
 
     git config --global dtcc-builder.autocrlf false
+
+## Remote development in VS Code
+
+On the left-side menu, go to Remote Explorer, on the SSH line press the + sign and add `username@develop.dtcc.chalmers.se` for the case of user `username` and the host being `develop.dtcc.chalmers.se`
+
+![image](https://user-images.githubusercontent.com/125367195/231126612-d6031bce-ca2d-4340-b0e5-9e728da57238.png)
+
+Then after `develop.dtcc.chalmers.se` appears in the list, click on the connect to current (right arrow) or new window (plus with folder) signs respectively.
+
+![image](https://user-images.githubusercontent.com/125367195/231126959-d1dc6498-576b-42aa-95ce-aff8df80c110.png)
+
+Then you can use `Open...` for opening a folder/file (eg /home/<username>/dtcc-builder) and `Terminal-> New terminal` to have a new terminal connected.
+
+
