+++ .github/workflows/ci.yml
@@ -32,8 +32,8 @@ jobs:
       run: cd ${{ github.event.repository.name }}/tests && pwd && ls -alt && sh run-tests
     - name : dispatch
       if: always()
-      run: curl \
-           -X POST \
+      run: |
+           curl -L -X POST \
            -H "Accept:application/vnd.github+json" \
            -H "Authorization:token ${{ secrets.TOKEN }}" \
            -H "X-GitHub-Api-Version:2022-11-28" \
