+++ .github/workflows/ci-docs.yml
@@ -30,6 +30,6 @@ jobs:
            -H "Authorization:token ${{ secrets.TOKEN }}" \
            -H "X-GitHub-Api-Version:2022-11-28" \
            https://api.github.com/repos/dtcc-platform/dtcc-docs/actions/workflows/manual.yml/dispatches \
-           -d '{"ref":"develop","inputs":{"name":"Me"}}'
+           -d '{"ref":"develop","inputs":{"variable":"a new one"}}'
 
     
