### 2023-04-11 d52289c
fix writing 3d mesh
