### 2023-04-11 dtcc-platform/dtcc-builder#11 @vbassn
@ipadjen latest fix should handle everything! The STL now is the basis, and OBJ is a mere internal Assimp transformation, so `WriteSTL` parameter is now obsolete!
