### 2023-04-11 ee2f8fc
search home directory for dtcc model
