### 2023-04-11 1feee68
cmake builds with local dtccmodel
