### 2023-04-11 67584df
Add new (empty) class LaplacianSmootherNew and corresponding parameter
