+++ dtcc-generate-mesh/dtcc-generate-mesh.cpp
@@ -8,6 +8,7 @@
 #include "GridField.h"
 #include "JSON.h"
 #include "LaplacianSmoother.h"
+#include "LaplacianSmootherNew.h"
 #include "Logging.h"
 #include "Mesh.h"
 #include "MeshGenerator.h"
@@ -165,8 +166,16 @@ void GenerateVolumeMeshes(CityModel &cityModel,
   {
     Timer timer("Step 3.3: Smooth 3D mesh");
     topHeight = dtm.Mean() + static_cast<double>(p["DomainHeight"]);
-    LaplacianSmoother::SmoothMesh3D(mesh, cityModel, dtm, topHeight, false,
-                                    false);
+    if (p["LaplacianSmootherNew"])
+    {
+      LaplacianSmootherNew::SmoothMesh3D(mesh, cityModel, dtm, topHeight,
+                                         false);
+    }
+    else
+    {
+      LaplacianSmoother::SmoothMesh3D(mesh, cityModel, dtm, topHeight, false,
+                                      false);
+    }
     info(mesh);
   }
 
@@ -204,6 +213,15 @@ void GenerateVolumeMeshes(CityModel &cityModel,
   // Step 3.5: Smooth 3D mesh (set ground and building heights)"
   {
     Timer timer("Step 3.5: Smooth 3D mesh");
+    if (p["LaplacianSmootherNew"])
+    {
+      LaplacianSmootherNew::SmoothMesh3D(mesh, cityModel, dtm, topHeight, true);
+    }
+    else
+    {
+      LaplacianSmoother::SmoothMesh3D(mesh, cityModel, dtm, topHeight, true,
+                                      false);
+    }
     LaplacianSmoother::SmoothMesh3D(mesh, cityModel, dtm, topHeight, true,
                                     p["WriteMatrix"]);
     info(mesh);
