+++ src/dtcc_builder/scripts/dtcc_build.py
@@ -16,13 +16,19 @@ from google.protobuf.json_format import MessageToJson
 import dtcc_io as io
 import dtcc_builder as builder
 
+
 def camel_to_kebab(name):
     name = re.sub("(.)([A-Z][a-z]+)", r"\1-\2", name)
     return re.sub("([a-z0-9])([A-Z])", r"\1-\2", name).lower()
 
 
 def create_parameters_options(parser):
-    type_map = {"str": str, "float": float, "int": int, "bool": bool}
+    type_map = {
+        "str": str,
+        "float": float,
+        "int": int,
+        "bool": bool,
+    }
 
     name_translator = {}
     p = builder.load_parameters()
@@ -32,11 +38,10 @@ def create_parameters_options(parser):
         name_translator[kebab_name] = k
         val_type = type(v).__name__
         if val_type == "bool":
-            parser.add_argument(f"--{kebab_name}",
-                                default=v, action="store_true")
+            parser.add_argument(f"--{kebab_name}", default=v, action="store_true")
         else:
             parser.add_argument(
-                f"--{kebab_name}", default=None, type=type_map[val_type]
+                f"--{kebab_name}", default=None, type=type_map.get(val_type, str)
             )
     return parser, name_translator
 
@@ -71,42 +76,48 @@ def get_project_paths(path):
 
     return (parameters_file, project_path)
 
-def run(p, citymodel_only, mesh_only):
 
+def run(p, citymodel_only, mesh_only):
     print(p)
-    #cm, dtm = None, None
-    cm, dtm = builder.build_citymodel(p["DataDirectory"] / p["BuildingsFileName"],p["PointCloudDirectory"],p )
+    # cm, dtm = None, None
+    cm, dtm = builder.build_citymodel(
+        p["DataDirectory"] / p["BuildingsFileName"], p["PointCloudDirectory"], p
+    )
     if p["WriteJSON"]:
-        with open(p["OutputDirectory"]/ "CityModel.json", "w") as dst:
+        with open(p["OutputDirectory"] / "CityModel.json", "w") as dst:
             dst.write(MessageToJson(cm))
     if p["WriteProtobuf"]:
-        with open(p["OutputDirectory"]/ "CityModel.pb", "wb") as dst:
+        with open(p["OutputDirectory"] / "CityModel.pb", "wb") as dst:
             dst.write(cm.SerializeToString())
-    io.save_citymodel(cm, p["OutputDirectory"] / "CityModel.shp",)
+    io.save_citymodel(
+        cm,
+        p["OutputDirectory"] / "CityModel.shp",
+    )
     if not citymodel_only:
-        volume_mesh, surface_mesh = builder.build_mesh(p["DataDirectory"] / p["BuildingsFileName"],p["PointCloudDirectory"],p )
+        volume_mesh, surface_mesh = builder.build_mesh(
+            p["DataDirectory"] / p["BuildingsFileName"], p["PointCloudDirectory"], p
+        )
 
         if p["WriteJSON"]:
-            with open(p["OutputDirectory"]/ "CitySurface.json", "w"):
+            with open(p["OutputDirectory"] / "CitySurface.json", "w"):
                 MessageToJson(surface_mesh)
-            with open(p["OutputDirectory"]/ "CityMesh.json", "w"):
+            with open(p["OutputDirectory"] / "CityMesh.json", "w"):
                 MessageToJson(volume_mesh)
-        
+
         if p["WriteProtobuf"]:
-            with open(p["OutputDirectory"]/ "CitySurface.pb", "wb") as dst:
-                dst.write(surface_mesh.SerilaizeToString())
-            with open(p["OutputDirectory"]/ "CityMesh.pb", "wb") as dst:
[…]
