+++ src/dtcc_builder/Parameters.py
@@ -54,15 +54,18 @@ _parameters["OutputDirectory"] = ""
 
 def load_parameters(file_path=None, project_path="."):
     global _parameters
-    #print(file_path)
+    # print(file_path)
     if file_path is None:
-        p = _parameters
+        p = _parameters.copy()
+        p = set_directories(p, project_path)
     else:
         file_path = Path(file_path)
         if file_path.is_dir():
             file_path = file_path / "Parameters.json"
         if not file_path.exists():
-            print(f"Parameters file {file_path} does not exist, using default parameters")
+            print(
+                f"Parameters file {file_path} does not exist, using default parameters"
+            )
             p = _parameters
         else:
             with open(file_path) as src:
@@ -70,6 +73,7 @@ def load_parameters(file_path=None, project_path="."):
             _parameters.update(loaded_parameters)
             p = set_directories(_parameters, project_path)
     return p
+
 
 def set_directories(p, project_path):
     if p["DataDirectory"] == "":
@@ -85,7 +89,6 @@ def set_directories(p, project_path):
     else:
         p["PointCloudDirectory"] = Path(p["PointCloudDirectory"])
         if not p["PointCloudDirectory"].is_absolute():
-            p["PointCloudDirectory"] = p["DataDirectory"] / \
-                p["PointCloudDirectory"]
+            p["PointCloudDirectory"] = p["DataDirectory"] / p["PointCloudDirectory"]
 
     return p
