### 2023-04-11 43021be
add test for build citymodel
