+++ .github/workflows/ci-docs.yml
@@ -29,7 +29,7 @@ jobs:
            -H "Accept:application/vnd.github+json" \
            -H "Authorization:token ${{ secrets.TOKEN }}" \
            -H "X-GitHub-Api-Version:2022-11-28" \
-           https://api.github.com/repos/dtcc-platform/dtcc-docs/actions/workflows/manual.yml/dispatches \
+           https://api.github.com/repos/dtcc-platform/dtcc-docs/actions/workflows/ci.yml/dispatches \
            -d '{"ref":"develop","inputs":{"Variable":"a new one"}}'
 
     
