+++ src/dtcc_builder/__init__.py
@@ -1,4 +1,3 @@
-
 from dtcc_builder.CityModel import CityModel
 from dtcc_builder.ElevationModel import ElevationModel
 from dtcc_builder.PointCloud import PointCloud
@@ -8,4 +7,4 @@ from dtcc_builder import _pybuilder
 
 from dtcc_builder.Parameters import load_parameters
 
-from dtcc_builder.Build import build_citymodel, build_surface_meshes, build_mesh
+from dtcc_builder.Build import build_citymodel, build_surface_meshes, build_volume_mesh
