+++ src/dtcc_builder/__init__.py
@@ -1,7 +1,6 @@
 from dtcc_builder.CityModel import CityModel
 from dtcc_builder.ElevationModel import ElevationModel
 from dtcc_builder.PointCloud import PointCloud
-from dtcc_builder.PointCloud import PointCloud
 from dtcc_builder import Meshing
 from dtcc_builder import _pybuilder
 
