+++ tests/python/test_pointcloud.py
@@ -6,34 +6,34 @@ sys.path.append(str((Path(__file__).parent / "..").resolve()))
 import unittest
 
 import dtcc_io as io
-from dtcc_builder import _pybuilder
-from dtcc_builder import Parameters
+import dtcc_builder as builder
 
-from dtcc_builder import PointCloud
+from dtcc_builder import _pybuilder
+
 
 data_dir = (Path(__file__).parent / "../data").resolve()
-p = Parameters.load_parameters(data_dir / "MinimalCase" / "Parameters.json")
+p = builder.Parameters.load_parameters(data_dir / "MinimalCase" / "Parameters.json")
 
 
 class TestLoadPointCloud(unittest.TestCase):
     def test_load_dir(self):
-        pc = PointCloud(data_dir / "MinimalCase")
+        pc = builder.PointCloud(data_dir / "MinimalCase")
         self.assertIsInstance(pc._builder_pc, _pybuilder.PointCloud)
 
     def test_load_file(self):
-        pc = PointCloud(data_dir / "MinimalCase" / "pointcloud.las")
+        pc = builder.PointCloud(data_dir / "MinimalCase" / "pointcloud.las")
         self.assertIsInstance(pc._builder_pc, _pybuilder.PointCloud)
         self.assertEqual(len(pc), 8148)
 
     def test_load_file_bounded(self):
-        pc = PointCloud(
+        pc = builder.PointCloud(
             data_dir / "MinimalCase" / "pointcloud.las", bounds=(-2, -2, 0, 0)
         )
         self.assertIsInstance(pc._builder_pc, _pybuilder.PointCloud)
         self.assertEqual(len(pc), 64)
 
     def test_load_dir_bounded(self):
-        pc = PointCloud(data_dir / "MinimalCase", bounds=(-2, -2, 0, 0))
+        pc = builder.PointCloud(data_dir / "MinimalCase", bounds=(-2, -2, 0, 0))
         self.assertIsInstance(pc._builder_pc, _pybuilder.PointCloud)
         self.assertEqual(len(pc), 64)
 
@@ -47,20 +47,20 @@ class TestLoadPointCloud(unittest.TestCase):
 
 class TestFilters(unittest.TestCase):
     def test_remove_outliers(self):
-        pc = PointCloud(data_dir / "MinimalCase" / "pointcloud.las")
+        pc = builder.PointCloud(data_dir / "MinimalCase" / "pointcloud.las")
         pc.remove_global_outliers(1)
         self.assertIsInstance(pc._builder_pc, _pybuilder.PointCloud)
         self.assertEqual(len(pc), 8148 - 1183)
 
     def test_remove_vegetation(self):
-        pc = PointCloud(data_dir / "MinimalCase" / "pointcloud.las")
+        pc = builder.PointCloud(data_dir / "MinimalCase" / "pointcloud.las")
         pc.vegetation_filter()
         self.assertIsInstance(pc._builder_pc, _pybuilder.PointCloud)
 
 
 class TestConvert(unittest.TestCase):
     def test_convert2numpy(self):
-        pc = PointCloud(data_dir / "MinimalCase" / "pointcloud.las")
+        pc = builder.PointCloud(data_dir / "MinimalCase" / "pointcloud.las")
         pts_array = pc.points
         l, d = pts_array.shape
         self.assertEqual(l, 8148)
@@ -70,24 +70,23 @@ class TestConvert(unittest.TestCase):
         pb_pc = str(data_dir / "MinimalCase" / "pointcloud.las.pb")
         with open(pb_pc, "rb") as src:
             pb_string = src.read()
-        pc = PointCloud()
+        pc = builder.PointCloud()
[…]
