### 2023-04-12 e72d417
Made some initial changes to LaplacianSmootherNew.h
