### 2023-04-12 dtcc-platform/dtcc-builder#12 @anderslogg
This is a known issue that we are planning to fix in an upcoming release based off the `develop` branch of dtcc-builder which has come a long way since the original submission of the JOSS article/branch.

Is it critical that this issue is fixed for the JOSS submission? @ipadjen
