### 2023-04-13 87031b6
fix bug parsing bool values from Parameters.json
