+++ src/dtcc_builder/scripts/dtcc_build.py
@@ -38,6 +38,7 @@ def create_parameters_options(parser):
         name_translator[kebab_name] = k
         val_type = type(v).__name__
         if val_type == "bool":
+            # print(f"bool arg: {kebab_name}")
             parser.add_argument(f"--{kebab_name}", default=v, action="store_true")
         else:
             parser.add_argument(
@@ -50,12 +51,16 @@ def update_parameters_from_options(p, args, name_translator):
     for k, v in p.items():
         print(k)
         snake_name = name_translator.get(k)
+        print(snake_name)
         if snake_name is None:
             continue
         snake_name = name_translator[k].replace("-", "_")
+        print(snake_name)
         parser_val = getattr(args, snake_name)
+        print(parser_val)
         if parser_val is not None:
             p[k] = parser_val
+        print("---")
     return p
 
 
@@ -129,7 +134,7 @@ def main():
 
     parser.add_argument("projectpath", nargs="?", default=os.getcwd())
 
-    parser, name_translator = create_parameters_options(parser)
+    # parser, name_translator = create_parameters_options(parser)
 
     args = parser.parse_args()
 
@@ -141,5 +146,5 @@ def main():
     else:
         parameters = builder.Parameters.load_parameters(parameters_file, project_path)
 
-    parameters = update_parameters_from_options(parameters, args, name_translator)
+    # parameters = update_parameters_from_options(parameters, args, name_translator)
     run(parameters, args.citymodel_only, args.mesh_only)
