### 2023-04-13 731d903
add --foo and --no-foo for boolean options
