+++ pyproject.toml
@@ -4,26 +4,26 @@ build-backend = "scikit_build_core.build"
 
 [project]
 name = "dtcc-builder"
-version = "0.5.1"
-description="DTCC Builder"
+version = "0.5.2"
+description = "DTCC Builder"
 readme = "README.md"
 authors = [
   { name = "Anders Logg", email = "logg@chalmers.se" },
   { name = "Vasilis Naserentin", email = "vasilis.naserentin@chalmers.se" },
-  { name = "Dag Wästberg", email = "dwastberg@gmail.com" }
+  { name = "Dag Wästberg", email = "dwastberg@gmail.com" },
 ]
 requires-python = ">=3.8"
 classifiers = [
   "Intended Audience :: Science/Research",
   "Programming Language :: Python :: 3",
   "Programming Language :: Python :: C++",
-  "License :: OSI Approved :: MIT License"
+  "License :: OSI Approved :: MIT License",
 ]
 dependencies = [
- "pybind11 >= 2.10.0, < 2.11.0",
- "numpy",
- "dtcc-model@git+https://github.com/dtcc-platform/dtcc-model.git@develop",
- "dtcc-io@git+https://github.com/dtcc-platform/dtcc-io.git@develop"
+  "pybind11 >= 2.10.0, < 2.11.0",
+  "numpy",
+  "dtcc-model@git+https://github.com/dtcc-platform/dtcc-model.git@develop",
+  "dtcc-io@git+https://github.com/dtcc-platform/dtcc-io.git@develop",
 ]
 
 [project.optional-dependencies]
