### 2023-04-14 #4928 @ipadjen
Hi @crvernon! I kindly asked authors to provide proof of their functional claims which I cannot confirm by simply running the code (e.g. show the results in the paper or somewhere in the repo). More here: https://github.com/dtcc-platform/dtcc-builder/issues/14.

Another issue that is currently open (https://github.com/dtcc-platform/dtcc-builder/issues/12) is not publication critical.
