### 2023-04-17 dtcc-platform/dtcc-builder#12 @anderslogg
Yes let's keep it open since we plan to fix this (in a future release).
