### 2023-04-17 dtcc-platform/dtcc-builder#14 @anderslogg
How much of a "real-world" simulation is required?

We can easily add a demo for solving a PDE on the generated meshes to the repository but it may not be a "real-world" case with proper boundary conditions, relevant parameters etc. Such simulations are in progress but require substantial effort to implement and validate. We could add simple demos such as the Stokes equations as a simple model of wind/fluid flow or the Poisson equation as a simple model of electromagnetic fields.
