+++ src/cpp/builder_bindings.cpp
@@ -1,0 +1,197 @@
+#include <pybind11/numpy.h>
+#include <pybind11/pybind11.h>
+
+#include <pybind11/stl.h>
+
+#include "CityModelGenerator.h"
+#include "Point.h"
+#include "PointCloud.h"
+#include "PointCloudMethods.h"
+#include "Polygon.h"
+#include "datamodel/Building.h"
+#include "datamodel/CityModel.h"
+
+namespace py = pybind11;
+
+namespace DTCC_BUILDER
+{
+CityModel createBuilderCityModel(py::list footprints,
+                                 py::list uuids,
+                                 py::list heights,
+                                 py::list ground_levels,
+                                 py::tuple origin)
+{
+  CityModel cityModel;
+
+  cityModel.Origin =
+      Point2D(origin[0].cast<double>(), origin[1].cast<double>());
+  size_t num_buildings = footprints.size();
+  for (size_t i = 0; i < num_buildings; i++)
+  {
+    auto footprint = footprints[i].cast<py::list>();
+    auto uuid = uuids[i].cast<std::string>();
+    auto height = heights[i].cast<double>();
+    auto ground_level = ground_levels[i].cast<double>();
+    Polygon poly;
+    for (size_t j = 0; j < footprint.size(); j++)
+    {
+      auto pt = footprint[j].cast<py::tuple>();
+      poly.Vertices.push_back(
+          Point2D(pt[0].cast<double>(), pt[1].cast<double>()));
+    }
+    Building building;
+    building.Footprint = poly;
+    building.UUID = uuid;
+    building.Height = height;
+    building.GroundHeight = ground_level;
+    cityModel.Buildings.push_back(building);
+  }
+  return cityModel;
+}
+
+PointCloud createBuilderPointCloud(py::array_t<double> pts,
+                                   py::array_t<uint8_t> cls,
+                                   py::array_t<uint8_t> retNumber,
+                                   py::array_t<uint8_t> numReturns)
+{
+  auto pts_r = pts.unchecked<2>();
+  auto cls_r = cls.unchecked<1>();
+  auto retNumber_r = retNumber.unchecked<1>();
+  auto numReturns_r = numReturns.unchecked<1>();
+
+  size_t pt_count = pts_r.shape(0);
+
+  bool hasClassification = cls_r.size() == pt_count;
+  bool hasReturnNumber = retNumber_r.size() == pt_count;
+  bool hasNumberOfReturns = numReturns_r.size() == pt_count;
+
+  PointCloud pointCloud;
+  for (size_t i = 0; i < pt_count; i++)
+  {
+    pointCloud.Points.push_back(Point3D(pts_r(i, 0), pts_r(i, 1), pts_r(i, 2)));
+    if (hasClassification)
+      pointCloud.Classifications.push_back(cls_r(i));
+    else
+      pointCloud.Classifications.push_back(1);
+    if (hasReturnNumber)
+      pointCloud.ScanFlags.push_back(
+          PointCloudProcessor::packScanFlag(retNumber_r(i), numReturns_r(i)));
+  }
+
+  pointCloud.BuildHasClassifications();
+  pointCloud.CalculateBoundingBox();
+  return pointCloud;
+}
+
+PointCloud removeGlobalOutliers(PointCloud &pointCloud, double outlierMargin)
+{
+  PointCloudProcessor::RemoveOutliers(pointCloud, outlierMargin);
+  return pointCloud;
+}
+
+PointCloud removeVegetation(PointCloud &pointCloud)
+{
+  PointCloudProcessor::NaiveVegetationFilter(pointCloud);
[…]
