### 2023-04-18 dtcc-platform/dtcc-builder#14 @ipadjen
Just add something that supports the statement from your summary. I think there's also no need for me to wait on you to do that, we can wrap up the review if you agree with that.
