+++ sandbox/smoothing-2023/include/boundaryConditions.hpp
@@ -9,17 +9,79 @@
 #include <string>
 #include <vector>
 
-// #include "GridField.h"
+#include "GridField.h"
 // #include "LaplacianSmoother.h"
-//#include "Mesh.h"
+// #include "Mesh.h"
 // #include "datamodel/CityModel.h"
+
+#include "stiffnessMatrix.hpp"
 
 using namespace DTCC;
 
-void getVerticeMarkers(const Mesh3D &mesh, int *vMarkers)
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+// Boundary Conditions Class (WIP)
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+class BoundaryConditions
 {
-  const size_t nC = mesh.Cells.size();
-  const size_t nV = mesh.Vertices.size();
+private:
+  Mesh3D &_mesh;
+
+  const CityModel &_citymodel;
+
+  const GridField2D &_dtm;
+
+  const double topHeight;
+
+  const bool fixBuildings;
+
+public:
+  std::vector<int> vMarkers;
+
+  std::vector<double> values;
+
+  BoundaryConditions(Mesh3D &mesh,
+                     const CityModel &cityModel,
+                     const GridField2D &dtm,
+                     const double topHeight,
+                     const bool fixBuildings);
+
+  ~BoundaryConditions();
+
+  void computeVerticeMarkers();
+
+  void computeBoundaryValues();
+
+  void apply(std::vector<double> &b);
+
+  void apply(stiffnessMatrix &A);
+};
+
+BoundaryConditions::BoundaryConditions(Mesh3D &mesh,
+                                       const CityModel &cityModel,
+                                       const GridField2D &dtm,
+                                       const double topHeight,
+                                       const bool fixBuildings)
+    : _mesh(mesh), _citymodel(cityModel), _dtm(dtm), topHeight(topHeight),
+      vMarkers(mesh.Vertices.size(), 0), values(mesh.Vertices.size(), 0.0),
+      fixBuildings(fixBuildings)
+{
+  const size_t nC = _mesh.Cells.size();
+  const size_t nV = _mesh.Vertices.size();
+
+  info("Translating Cell Markers to Vertice Markers");
+  computeVerticeMarkers();
+
+  info("Computing Boundary Values");
+  computeBoundaryValues();
+}
+
+BoundaryConditions::~BoundaryConditions() {}
+
+void BoundaryConditions::computeVerticeMarkers()
+{
+  const size_t nC = _mesh.Cells.size();
+  const size_t nV = _mesh.Vertices.size();
 
   for (size_t i = 0; i < nV; i++)
   {
@@ -35,22 +97,22 @@ void getVerticeMarkers(const Mesh3D &mesh, int *vMarkers)
   for (size_t cn = 0; cn < nC; cn++)
   {
     // Initializing Global Index for each cell
-    I[0] = mesh.Cells[cn].v0;
-    I[1] = mesh.Cells[cn].v1;
-    I[2] = mesh.Cells[cn].v2;
-    I[3] = mesh.Cells[cn].v3;
+    I[0] = _mesh.Cells[cn].v0;
+    I[1] = _mesh.Cells[cn].v1;
+    I[2] = _mesh.Cells[cn].v2;
+    I[3] = _mesh.Cells[cn].v3;
 
-    const double z_mean = (mesh.Vertices[I[0]].z + mesh.Vertices[I[1]].z +
-                           mesh.Vertices[I[2]].z + mesh.Vertices[I[3]].z) /
[…]
