+++ sandbox/smoothing-2023/include/assembled.hpp
@@ -10,9 +10,12 @@
 #include <string>
 #include <vector>
 
+#include "GridField.h"
 #include "Mesh.h"
+#include "Timer.h"
 #include "datamodel/CityModel.h"
 
+#include "boundaryConditions.hpp"
 #include "error.hpp"
 #include "sparse.hpp"
 #include "stiffnessMatrix.hpp"
@@ -93,7 +96,7 @@ void assembled_GaussSeidel(const Mesh3D &mesh,
 
   for (size_t cn = 0; cn < nC; cn++)
   {
-    // Initializing Global Index for each cell
+    // Global Index for each cell
     I[0] = mesh.Cells[cn].v0;
     I[1] = mesh.Cells[cn].v1;
     I[2] = mesh.Cells[cn].v2;
@@ -153,5 +156,157 @@ void assembled_GaussSeidel(const Mesh3D &mesh,
 
   delete[] isBoundary, b, d, u_new, u_old, AK;
 }
+
+// Test function: Naive BC Application
+void checkBoundaryPoints(bool *isBoundary,
+                         double *boundaryValues,
+                         const Mesh3D &m)
+{
+  const size_t nV = m.Vertices.size();
+  const double epsilon = 1;
+  const double boundary_line1 = 37.0;
+  const double boundary_line2 = 100.0;
+
+  for (size_t i = 0; i < nV; i++)
+  {
+    if ((abs(m.Vertices[i].z - boundary_line2) <
+         epsilon)) //|| (abs(m->Vertices[i].z - boundary_line2) < epsilon))
+    {
+      isBoundary[i] = true;
+      boundaryValues[i] = 3;
+    }
+  }
+  return;
+}
+
+// Laplacian Smoother using Sparse format Stiffness Matrix
+// for Gauss-Seidel
+class SparseSmoother
+{
+public:
+  static void SmoothMesh3D(Mesh3D &mesh3D,
+                           const CityModel &cityModel,
+                           const GridField2D &dem,
+                           double topHeight,
+                           bool fixBuildings)
+  {
+    info("LaplacianSmoother: Smoothing mesh (Laplacian smoothing NEW)...");
+    Timer timer("SmoothMesh3DNew: Sparse");
+
+    const size_t nV = mesh3D.Vertices.size();
+
+    std::vector<double> b(nV, 0);
+
+    BoundaryConditions bc(mesh3D, cityModel, dem, topHeight, fixBuildings);
+    bc.apply(b);
+
+    stiffnessMatrix AK(mesh3D);
+
+    COO_array A_coo(nV, nV);
+    assemble(AK, bc.vMarkers, A_coo);
+
+    CSR_array A(&A_coo);
+
+    std::vector<double> u = b;
+
+    GaussSeidel(mesh3D, A, b, u);
+
+    // Update mesh coordinates
+    for (std::size_t i = 0; i < mesh3D.Vertices.size(); i++)
+      mesh3D.Vertices[i].z += u[i];
+  }
+
+  static void assemble(const stiffnessMatrix &A,
+                       const std::vector<int> &VerticeMarkers,
+                       COO_array &assembled_A)
+  {
+    info("Assembling Stiffness Matrix ");
+
+    size_t nC = A._mesh.Cells.size();
+    size_t nV = A._mesh.Vertices.size();
+
+    std::array<uint, 4> I;
+
+    for (size_t i = 0; i < nV; i++)
+    {
+      if (VerticeMarkers[i] > -4)
+      {
+        assembled_A.add(i, i, 1);
+      }
+    }
+
+    std::cout << "Inserterd Boundary Vertices" << std::endl;
+
[…]
