+++ sandbox/benchmarks/benchmark_roofpoint_extract.py
@@ -1,5 +1,5 @@
 import dtcc_io as io
-import dtcc_builder.Build
+import dtcc_builder.build
 
 cm = io.load_citymodel(
     "/home/cit-wastberg/dtcc-platform/dtcc-builder/data/HelsingborgResidential2022/PropertyMap.shp"
@@ -8,4 +8,4 @@ pc = io.load_pointcloud(
     "/home/cit-wastberg/dtcc-platform/dtcc-builder/data/HelsingborgResidential2022/PointCloud.las"
 )
 
-cm = dtcc_builder.Build.extract_roofpoints(cm, pc)
+cm = dtcc_builder.build.extract_buildingpoints(cm, pc)
