+++ src/dtcc_builder/__init__.py
@@ -3,7 +3,7 @@
 # from dtcc_builder.PointCloud import PointCloud
 # from dtcc_builder import Meshing
 # from dtcc_builder import _pybuilder
-
-# from dtcc_builder.Parameters import load_parameters
+from dtcc_builder import build
+from dtcc_builder.Parameters import load_parameters
 
 # from dtcc_builder.Build import build_citymodel, build_surface_meshes, build_volume_mesh
