### 2023-04-24 dtcc-platform/dtcc-builder#14 @anderslogg
@ipadjen Alright, I'll add a simple PDE demo and point to it in the documentation. Easy to add and should be done this week.
