### 2023-04-24 2166475
Write mesh to file in FEniCS XML format
