+++ src/cpp/builder_bindings.cpp
@@ -6,6 +6,10 @@
 #include "CityModelGenerator.h"
 #include "ElevationModelGenerator.h"
 #include "GridField.h"
+#include "LaplacianSmoother.h"
+#include "Mesh.h"
+#include "MeshGenerator.h"
+#include "MeshProcessor.h"
 #include "Point.h"
 #include "PointCloud.h"
 #include "PointCloudMethods.h"
@@ -86,6 +90,37 @@ PointCloud createBuilderPointCloud(py::array_t<double> pts,
   return pointCloud;
 }
 
+GridField2D createBuilderGridField(py::array_t<double> data,
+                                   py::tuple bounds,
+                                   size_t XStep,
+                                   size_t YStep,
+                                   double XSize,
+                                   double YSize)
+{
+  GridField2D gridField;
+  double px = bounds[0].cast<double>();
+  double py = bounds[1].cast<double>();
+  double qx = bounds[2].cast<double>();
+  double qy = bounds[3].cast<double>();
+  auto bbox = BoundingBox2D(Point2D(px, py), Point2D(qx, qy));
+
+  gridField.Grid.BoundingBox = bbox;
+  gridField.Grid.XStep = XStep;
+  gridField.Grid.YStep = YStep;
+  gridField.Grid.XSize = XSize;
+  gridField.Grid.YSize = YSize;
+
+  auto data_r = data.unchecked<1>();
+  size_t data_count = data_r.size();
+
+  for (size_t i = 0; i < data_count; i++)
+  {
+    gridField.Values.push_back(data_r(i));
+  }
+
+  return gridField;
+}
+
 PointCloud removeGlobalOutliers(PointCloud &pointCloud, double outlierMargin)
 {
   PointCloudProcessor::RemoveOutliers(pointCloud, outlierMargin);
@@ -140,6 +175,108 @@ GridField2D SmoothElevation(GridField2D &dem, size_t numSmoothings)
   return dem;
 }
 
+CityModel SimplifyCityModel(CityModel &cityModel,
+                            py::tuple bounds,
+                            double minimalBuildingDistance,
+                            double minimalVertexDistance)
+{
+  double px = bounds[0].cast<double>();
+  double py = bounds[1].cast<double>();
+  double qx = bounds[2].cast<double>();
+  double qy = bounds[3].cast<double>();
+  auto bbox = BoundingBox2D(Point2D(px, py), Point2D(qx, qy));
+  CityModelGenerator::SimplifyCityModel(
+      cityModel, bbox, minimalBuildingDistance, minimalVertexDistance);
+  return cityModel;
+}
+
+CityModel CleanCityModel(CityModel &cityModel, double minVertDistance)
+{
+  CityModelGenerator::CleanCityModel(cityModel, minVertDistance);
+  return cityModel;
+}
+
+// Meshing
+
+Mesh2D
+GenerateMesh2D(const CityModel &cityModel, py::tuple bounds, double resolution)
+{
+  Mesh2D mesh;
+  double px = bounds[0].cast<double>();
+  double py = bounds[1].cast<double>();
+  double qx = bounds[2].cast<double>();
+  double qy = bounds[3].cast<double>();
+  auto bbox = BoundingBox2D(Point2D(px, py), Point2D(qx, qy));
+
[…]
