+++ src/dtcc_builder/build.py
@@ -10,6 +10,8 @@ import numpy as np
 from affine import Affine
 from typing import Union, Tuple
 
+import rasterio.transform
+
 from pypoints2grid import points2grid
 
 from time import time
@@ -65,16 +67,21 @@ def generate_dem(
 ) -> model.Raster:
     ground_point_idx = np.where(np.isin(pointcloud.classification, [2, 9]))[0]
     ground_points = pointcloud.points[ground_point_idx]
-    if isinstance(bounds, model.Bounds):
-        bounds = bounds.tuple
-    dem = points2grid(ground_points, cell_size, bounds, window_size=window_size)
 
+    print(f"generating dem with bounds {bounds.tuple}")
+    dem = points2grid(ground_points, cell_size, bounds.tuple, window_size=window_size)
+
+    print("dem shape: ", dem.shape)
+    print("cell size: ", cell_size)
     dem_raster = model.Raster()
     dem_raster.data = dem
     dem_raster.nodata = 0
-    dem_raster.georef = Affine.translation(*bounds[:2]) * Affine.scale(
-        cell_size, cell_size
+    dem_raster.georef = rasterio.transform.from_origin(
+        bounds.west, bounds.north, cell_size, cell_size
     )
+
+    print(f"generated dem with bounds {dem_raster.bounds}")
+    print(f"generated dem with georef \n{dem_raster.georef}")
     io.process.raster.fill_holes(dem_raster)
     return dem_raster
 
@@ -150,9 +157,9 @@ def build_mesh(
     city_model: model.CityModel,
     mesh_resolution,
     domain_height,
-    num_layers,
     min_building_dist,
     min_vertex_dist,
+    debug=False,
 ) -> Tuple[model.VolumeMesh, model.Mesh]:
     builder_cm = builder_datamodel.create_builder_citymodel(city_model)
     bounds = (
@@ -161,6 +168,8 @@ def build_mesh(
         city_model.bounds.xmax,
         city_model.bounds.ymax,
     )
+    print(f"Building mesh with bounds {bounds}")
+    print(f"dem bounds {city_model.terrain.bounds}")
     simple_cm = _pybuilder.SimplifyCityModel(
         builder_cm, bounds, min_building_dist, min_vertex_dist
     )
@@ -168,47 +177,53 @@ def build_mesh(
 
     builder_dem = builder_datamodel.raster_to_builder_gridfield(city_model.terrain)
 
+    # Step 3.1: Generate 2D mesh
     mesh_2D = _pybuilder.GenerateMesh2D(
         simple_cm,
         bounds,
         mesh_resolution,
     )
+
+    # Step 3.2: Generate 3D mesh (layer 3D mesh)
     mesh_3D = _pybuilder.GenerateMesh3D(mesh_2D, domain_height, mesh_resolution)
 
+    # Step 3.3: Smooth 3D mesh (set ground height)
     mesh_3D = _pybuilder.SmoothMesh3D(mesh_3D, simple_cm, builder_dem, False)
 
-    mesh_3D = _pybuilder.TrimMesh3D(mesh_3D, mesh_2D, simple_cm, num_layers)
+    # Step 3.4: Trim 3D mesh (remove building interiors)
+    mesh_3D = _pybuilder.TrimMesh3D(mesh_3D, mesh_2D, simple_cm)
 
+    # Step 3.5: Smooth 3D mesh (set ground and building heights)
     mesh_3D = _pybuilder.SmoothMesh3D(mesh_3D, simple_cm, builder_dem, True)
 
     surface_3d = _pybuilder.ExtractBoundary3D(mesh_3D)
 
+    # Step 3.6: Convert back to dtcc format
[…]
