+++ src/cpp/_pybind_builder.cpp
@@ -1,587 +1,0 @@
-#include <pybind11/numpy.h>
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
-
-#include "Protobuf.h"
-#include "datamodel/Building.h"
-#include "datamodel/CityModel.h"
-
-// Needs to come before JSON (nlohmann) include because of sloppy
-// namespacing in VTK (typedef detail)...
-// #include "VTK.h"
-
-// DTCC includes
-#include "CityModelGenerator.h"
-#include "ElevationModelGenerator.h"
-#include "GridField.h"
-#include "LaplacianSmoother.h"
-#include "Logging.h"
-#include "Mesh.h"
-#include "MeshGenerator.h"
-// #include "MeshIO.h"
-#include "MeshProcessor.h"
-#include "Point.h"
-#include "PointCloud.h"
-#include "Polygon.h"
-#include "Protobuf.h"
-#include "Surface.h"
-#include "VertexSmoother.h"
-
-namespace py = pybind11;
-
-namespace DTCC_BUILDER
-{
-
-// CityModel
-
-CityModel SetCityModelOrigin(CityModel &cityModel, py::tuple origin)
-{
-  double px = origin[0].cast<double>();
-  double py = origin[1].cast<double>();
-  Point2D o = Point2D(px, py);
-  cityModel.SetOrigin(o);
-
-  return cityModel;
-}
-
-CityModel CleanCityModel(CityModel &cityModel, double minVertDistance)
-{
-  CityModelGenerator::CleanCityModel(cityModel, minVertDistance);
-  return cityModel;
-}
-
-CityModel SimplifyCityModel(CityModel &cityModel,
-                            py::tuple bounds,
-                            double minimalBuildingDistance,
-                            double minimalVertexDistance)
-{
-  double px = bounds[0].cast<double>();
-  double py = bounds[1].cast<double>();
-  double qx = bounds[2].cast<double>();
-  double qy = bounds[3].cast<double>();
-  auto bbox = BoundingBox2D(Point2D(px, py), Point2D(qx, qy));
-  CityModelGenerator::SimplifyCityModel(
-      cityModel, bbox, minimalBuildingDistance, minimalVertexDistance);
-  return cityModel;
-}
-
-CityModel ExtractBuildingPoints(CityModel &cityModel,
-                                const PointCloud &pointCloud,
-                                double groundMargin,
-                                double groundOutlierMargin)
-{
-  CityModelGenerator::ExtractBuildingPoints(cityModel, pointCloud, groundMargin,
-                                            groundOutlierMargin);
-  return cityModel;
-}
-
-CityModel BuildingPointsRANSACOutlierRemover(CityModel &cityModel,
-                                             double outlier_margin,
-                                             size_t iterations)
-{
-  CityModelGenerator::BuildingPointsRANSACOutlierRemover(
-      cityModel, outlier_margin, iterations);
-  return cityModel;
-}
-
-CityModel BuildingPointsOutlierRemover(CityModel &cityModel,
-                                       size_t neighbors,
-                                       double outlier_margin)
-{
-  CityModelGenerator::BuildingPointsOutlierRemover(cityModel, neighbors,
-                                                   outlier_margin);
-  return cityModel;
-}
-
-CityModel ComputeBuildingHeights(CityModel &cityModel,
[…]
