+++ src/dtcc_builder/build.py
@@ -84,8 +84,6 @@ def generate_dem(
     print(f"generated dem with georef \n{dem_raster.georef}")
     dem_raster = io.process.raster.fill_holes(dem_raster)
 
-    io.save_raster(dem_raster, Path("harbor_dem.tif"))
-
     return dem_raster
 
 
@@ -197,7 +195,8 @@ def build_mesh(
     mesh_3D = _pybuilder.GenerateMesh3D(mesh_2D, domain_height, mesh_resolution)
     if debug:
         io.save_mesh(
-            builder_datamodel.builder_mesh3D_to_volume_mesh(mesh_3D), "mesh_step3.2.vtu"
+            builder_datamodel.builder_mesh3D_to_volume_mesh(mesh_3D),
+            "meshing_step3.2.vtu",
         )
 
     top_height = domain_height + city_model.terrain.data.mean()
@@ -207,20 +206,23 @@ def build_mesh(
     )
     if debug:
         io.save_mesh(
-            builder_datamodel.builder_mesh3D_to_volume_mesh(mesh_3D), "mesh_step3.3.vtu"
+            builder_datamodel.builder_mesh3D_to_volume_mesh(mesh_3D),
+            "meshing_step3.3.vtu",
         )
 
     # Step 3.4: Trim 3D mesh (remove building interiors)
     mesh_3D = _pybuilder.TrimMesh3D(mesh_3D, mesh_2D, simple_cm)
     if debug:
         io.save_mesh(
-            builder_datamodel.builder_mesh3D_to_volume_mesh(mesh_3D), "mesh_step3.4.vtu"
+            builder_datamodel.builder_mesh3D_to_volume_mesh(mesh_3D),
+            "meshing_step3.4.vtu",
         )
     # Step 3.5: Smooth 3D mesh (set ground and building heights)
     mesh_3D = _pybuilder.SmoothMesh3D(mesh_3D, simple_cm, builder_dem, top_height, True)
     if debug:
         io.save_mesh(
-            builder_datamodel.builder_mesh3D_to_volume_mesh(mesh_3D), "mesh_step3.5.vtu"
+            builder_datamodel.builder_mesh3D_to_volume_mesh(mesh_3D),
+            "meshing_step3.5.vtu",
         )
     surface_3d = _pybuilder.ExtractBoundary3D(mesh_3D)
 
