+++ src/dtcc_builder/CityModel.py
@@ -1,183 +1,0 @@
-from google.protobuf.json_format import MessageToJson
-
-from dtcc_builder import _pybuilder
-
-import numpy
-
-import dtcc_io as io
-from dtcc_builder import PointCloud, Parameters, ElevationModel
-import dtcc_model as model
-
-# from dtccpybuilder.PointCloud import PointCloud
-# from dtccpybuilder.ElevationModel import ElevationModel
-from typing import List, Tuple
-
-# from dtccpybuilder.Parameters import load_parameters
-
-
-class CityModel:
-    def __init__(
-        self, footprints_file=None, pb_string=None, parameters=None, bounds=None
-    ):
-        if parameters is None:
-            parameters = Parameters.load_parameters()
-        self.parameters = parameters
-        self.bounds = bounds
-        self._builder_cm = None
-        self.origin = (0, 0)
-        self.cleaned = False
-        self.simplified = False
-        self.extracted_points = False
-        self.calculated_heights = False
-        if footprints_file is not None:
-            cm_pb = io.load_citymodel(
-                footprints_file,
-                uuid_field=parameters["UUIDField"],
-                area_filter=self.parameters["MinBuildingSize"],
-                bounds=bounds,
-                min_edge_distance=self.parameters["MinBuildingDistance"],
-                return_serialized=True,
-            )
-            self.load_protobuf(cm_pb)
-        if pb_string is not None:
-            self.load_protobuf(pb_string)
-
-    def get_buildings(self):
-        if self._builder_cm is None:
-            return None
-        else:
-            return self._builder_cm.buildings
-
-    def set_buildings(self, b):
-        return None
-
-    buildings = property(get_buildings)
-
-    def generate_citymodel(self, shp_footprint_file):
-        """load shp file of building footprints"""
-        if self.bounds is None:
-            self.bounds = io.citymodel.building_bounds(
-                shp_footprint_file, self.parameters["DomainMargin"]
-            )
-        self._builder_cm = _pybuilder.GenerateCityModel(
-            str(shp_footprint_file),
-            self.bounds,
-            float(self.parameters["MinBuildingDistance"]),
-            float(self.parameters["MinBuildingSize"]),
-        )
-
-    def set_origin(self, origin: Tuple[float, float]):
-        """set the origin for the city model. Everything will be offset so that origin is at (0,0)"""
-        self._builder_cm = _pybuilder.SetCityModelOrigin(self._builder_cm, origin)
-        self.origin = origin
-
-    def clean_citymodel(self, min_vert_distance=None):
-        """fix any errors in the building polygons and simplify complex polygons"""
-        if min_vert_distance is None:
-            min_vert_distance = self.parameters["MinVertexDistance"]
-        self._builder_cm = _pybuilder.CleanCityModel(
-            self._builder_cm, min_vert_distance
-        )
-        self.cleaned = True
-
-    def simplify_citymodel(self, bounds):
-        self._builder_cm = _pybuilder.SimplifyCityModel(
-            self._builder_cm,
-            bounds,
-            self.parameters["MinBuildingDistance"],
-            self.parameters["MinVertexDistance"],
-        )
-        self.simplified = True
-
-    def extract_building_points(
-        self,
-        point_cloud: PointCloud,
-        ground_margins=None,
-        ground_outlier_margin=None,
-    ):
-        """Extract roof and ground points from the point cloud for each building footprint"""
-
-        if ground_margins is None:
-            ground_margins = self.parameters["GroundMargin"]
-        if ground_outlier_margin is None:
-            ground_outlier_margin = self.parameters["OutlierMargin"]
-
[…]
