### 2023-04-27 #4928 @crvernon
🚨 MIDWEEK RALLY 🚨

@ipadjen and @alogg - Could you please provide an update to how things are going here? As always, let me know if you have any questions!

Have a great day!
