### 2023-04-27 #4928 @anderslogg
@crvernon @ipadjen

I think we're only waiting for fixing this issue: https://github.com/dtcc-platform/dtcc-builder/issues/14

I have it fixed already locally but haven't had time to push it and add to the docs due to other pressing matters (teaching). Will fix soon...

![image](https://user-images.githubusercontent.com/125354137/234863603-36482a55-1461-4f47-8c09-4e6833e32602.png)
