+++ tests/python/test_builder_datamodel.py
@@ -22,14 +22,3 @@ class TestBuilderCityModel(unittest.TestCase):
         builder_cm = builder.builder_datamodel.create_builder_citymodel(cm)
         self.assertEqual(len(builder_cm.buildings), len(cm.buildings))
         self.assertEqual(len(builder_cm.buildings), 5)
-
-
-class TestBuilderMesh(unittest.TestCase):
-    def test_convert_mesh(self):
-        mesh = io.load_mesh(data_dir / ""cube.vtk")
-        builder_mesh = builder.builder_datamodel.create_builder_mesh(mesh)
-        
-        self.assertEqual(len(builder_mesh.vertices), len(mesh.vertices))
-        self.assertEqual(len(builder_mesh.vertices), 8)
-        self.assertEqual(len(builder_mesh.faces), len(mesh.faces))
-        self.assertEqual(len(builder_mesh.faces), 12)
