+++ doc/usage.md
@@ -234,3 +234,36 @@ by the `GroundSmoothing` parameter:
 experimental parameters may be added/removed during development. For
 a complete list of  parameters, refer to the parameter files
 `dtcc-generate-[citymodel,mesh].json` generated by running the demo.
+
+## Simulation demo
+
+To demonstrate the usefulness of the generated meshes for numerical
+simulation, a simple demo is provided in the `demo` directory.
+
+To run the simulation demo, first run the mesh generation demo:
+
+    ./dtcc-builder-demo
+
+and select `SimulationDemo2023`. This will generate a mesh using
+the same dataset as `HelsingborgResidential2022` using a smaller
+domain and slightly larger mesh size (in order to create a more
+lightweight mesh). In addition, the height of a couple of buildings
+are extended to create a more interesting geometry. The generated
+mesh will be saved to the `demo` directory in the file `mesh.xml`
+using the FEniCS/DOLFIN XML format.
+
+Next, run the simulation demo:
+
+    ./dtcc-simulation-demo
+
+The demo solves the Stokes equations on the generated mesh using
+a standard Taylor-Hood formulation using
+[FEniCS](https://www.fenicsproject.org).
+
+Output data is saved to the two files `velocity.pvd` and `pressure.pvd`
+(including the corresponding `.vtu` files) for visualization in Paraview.
+The image below shows the streamlines for the computed velocity field.
+
+![](images/simulation.png)
+*Velocity streamlines for the solution of the Stokes equations solved using the
+simulation demo.*
