### 2023-05-03 dtcc-platform/dtcc-builder#14 @anderslogg
@ipadjen I have now added a new demo to the repository (the script `demo/dtcc-simulation-demo`). Documentation has also been added on how to run the demo.
