### 2023-05-03 #4928 @crvernon
@alogg could you point me to where the following is satisfied in your repository?:

> Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

Thanks!
