### 2023-05-03 #4928 @crvernon
@anderslogg here are some other things that need to be addressed in your paper:
- [ ] In the "Summary" section, please define the acronym on the first use for DTCC.
- [ ] Lines 26-28.  The citations for previous work for 3D mesh generation for city modeling should be sited in a statement that either accompanies or precedes your statement that "DTCC Builder aims to solve these challenges..." by stating something along the lines of "Previous research on 3D mesh generation for cities demonstrated <content here> but do not fully account for <content here> (Ledoux et al, 2019; Ledoux et al., 2021; Paden et al., 2022).   DTCC Builder aims to solve these challenges by..."
- [ ] Line 37:  State that Gothenburg is in Sweden, (e.g. Gothenburg, Sweden)
- [ ] Lines 42-44.  Should read, `Building heights and ground height are incorporated through a PDE-based smoothing process as described by Naserentin & Logg (2022).` and then dump the last sentence in the paragraph.
- [ ] Line 49.  Remove the link to the GitLab pages as the code here no longer matches what you have represented in this review.
- [ ] See to the missing DOIs:

> MISSING DOIs
> - 10.3389/fbuil.2022.899332 may be a valid DOI for title: Towards Automatic Reconstruction of 3D City Models Tailored for Urban Flow Simulations
> - 10.1007/bfb0014497 may be a valid DOI for title: Triangle: Engineering a 2D Quality Mesh Generator and Delaunay Triangulator

Thanks!
