### 2023-05-04 680b87a
Fixing issues mentioned in https://github.com/openjournals/joss-reviews/issues/4928#issuecomment-1533250828
