### 2023-05-04 8511c46
Updated paper.md with more fixes regarding https://github.com/openjournals/joss-reviews/issues/4928#issuecomment-1533250828
