### 2023-05-04 44af3fa
Remove extra contributing section (now part of main README)
