+++ doc/contributing.md
@@ -1,13 +1,0 @@
-# Contributing
-
-Contributions to DTCC Builder are welcome and may come in the form of
-bug reports, feature requests, documentation, or code.
-
-To contribute, register a GitLab account and post an issue detailing
-your bug report, feature request or idea for writing documentation or
-code on the project page for
-[DTCC Builder](https://gitlab.com/dtcc-platform/dtcc-builder).
-
-Make sure to also check out the [developer notes](development.md) and
-note that any source code contributions need to be made open-source
-under the MIT License.
