### 2023-05-04 82bb473
Fixed pointy roof bug
