+++ sandbox/smoothing-2023/dtcc-sandbox.cpp
@@ -10,20 +10,94 @@
 // namespacing in VTK (typedef detail)...
 
 //  DTCC includes
-#include "include/JSON.hpp"
-// #include "JSON.h"
-#include "LaplacianSmoother.h"
-#include "Mesh.h"
-#include "Timer.h"
-#include "datamodel/CityModel.h"
 
+// #include "JSON.h"
+#include <cmath>
+#include <dolfin.h>
+#include <iostream>
+
+#include "LaplacianSmoother.h"
+
+// Sandbox Includes
+#include "include/JSON.hpp"
+// #include "include/assembled.hpp"
+// #include "include/fem.hpp"
 #include "include/assembled.hpp"
-#include "include/fem.hpp"
+#include "include/boundaryConditions.hpp"
 #include "include/sparse.hpp"
+#include "include/stiffnessMatrix.hpp"
 
 using namespace DTCC;
 
 void Help() { error("Usage: dtcc-sandbox Data_dir NumOfIterations"); }
+
+void checkMeshFenics(Mesh3D &mesh,
+                     CityModel &cityModel,
+                     const GridField2D &dem,
+                     double topHeight,
+                     bool fixBuildings,
+                     bool writeMat)
+{
+  info("Writing Stiffness Matrix A");
+
+  // Disable dof reordering
+  dolfin::parameters["reorder_dofs_serial"] = false;
+
+  LaplacianSmoother::SmoothMesh3D(mesh, cityModel, dem, topHeight, fixBuildings,
+                                  true);
+  return;
+}
+
+void checkMeshGS(Mesh3D &mesh,
+                 CityModel &cityModel,
+                 const GridField2D &dem,
+                 double topHeight,
+                 bool fixBuildings,
+                 bool writeMat)
+{
+  info("Writing Stiffness Matrix A: GS");
+
+  const std::size_t nV = mesh.Vertices.size();
+
+  stiffnessMatrix AK(mesh);
+
+  std::vector<double> b(mesh.Vertices.size(), 0);
+
+  BoundaryConditions bc(mesh, cityModel, dem, topHeight, fixBuildings);
+  bc.apply(b);
+  bc.apply(AK);
+
+  std::cout << "Assembled A ... " << std::endl;
+  std::vector<double> A = AK.assemble();
+
+  for (size_t i = 0; i < nV; i++)
+  {
+    if (bc.vMarkers[i] > -4)
+      A[nV * i + i] = 1;
+  }
+
+  std::stringstream ss;
+  for (size_t i = 0; i < nV; i++)
+  {
+    for (size_t j = 0; j < nV; j++)
+    {
+      if (A[i * nV + j] != 0)
+        ss << (i + 1) << " " << (j + 1) << " " << A[i * nV + j] << std::endl;
+    }
+  }
+
+  std::ofstream fs;
+  auto ff = "Matrix3_3.dat";
+  if (fixBuildings)
+  {
+    ff = "Matrix3_5.dat";
+  }
+  fs.open(ff);
+  fs << ss.rdbuf();
+  fs.close();
+
+  return;
+}
 
 int main(int argc, char *argv[])
 {
@@ -36,55 +110,28 @@ int main(int argc, char *argv[])
   std::string data_dir = argv[1];
   size_t max_iterations = atoi(argv[2]);
 
-  Mesh3D m;
-  CityModel cm;
-  GridField2D dtm;
+  Mesh3D mesh;
+  CityModel cityModel;
+  GridField2D dem;
 
   std::string mesh_dir = data_dir + "/CityMesh.json";
   std::string citymodel_dir = data_dir + "/CityModel.json";
[…]

[…]
