### 2023-05-04 8b1b11c
Fixed pointy roof bug
