+++ doc/paper-joss-2022/paper.bib
@@ -68,7 +68,7 @@
   year = {2022},
   journal = {Frontiers in Built Environment},
   volume = {8},
-  doi = {10.3389/fbuil.2022.899332}
+  doi = {10.3389/fbuil.2022.899332},
   issn = {2297-3362},
   abstract = {In the computational fluid dynamics simulation workflow, the geometry preparation step is often regarded as a tedious, time-consuming task. Many practitioners consider it one of the main bottlenecks in the simulation process. The more complex the geometry, the longer the necessary work, meaning this issue is amplified for urban flow simulations that cover large areas with complex building geometries. To address the issue of geometry preparation, we propose a workflow for automatically reconstructing simulation-ready 3D city models. The workflow combines 2D geographical datasets (e.g., cadastral data, topographic datasets) and aerial point cloud-based elevation data to reconstruct terrain, buildings, and imprint surface layers like water, low vegetation, and roads. Imprinted surface layers serve as different roughness surfaces for modeling the atmospheric boundary layer. Furthermore, the workflow is capable of automatically defining the influence region and domain size according to best practice guidelines. The resulting geometry aims to be error-free: without gaps, self-intersections, and non-manifold edges. The workflow was implemented into an open-source framework using modern, robust, and state-of-the-art libraries with the intent to be used for further developments. Our approach limits the geometry generation step to the order of hours (including input data retrieval and preparation), producing geometries that can be directly used for computational grid generation without additional preparation. The reconstruction done by the algorithm can last from a few seconds to a few minutes, depending on the size of the input data. We obtained and prepared the input data for our verification study in about 2~hours, while the reconstruction process lasted 1~minute. The unstructured computational meshes we created in an automatic mesh generator show satisfactory quality indicators and the subsequent numerical simulation exhibits good convergence behavior with the grid convergence index of observed variables less than 5\%.},
   file = {/Users/logg/Zotero/storage/YN5ST9RG/Pađen m. fl. - 2022 - Towards automatic reconstruction of 3D city models.pdf}
@@ -79,7 +79,7 @@
   booktitle = {Workshop on {{Applied Computational Geometry}}},
   author = {Shewchuk, Jonathan Richard},
   year = {1996},
-  doi = {10.1007/bfb0014497}
+  doi = {10.1007/bfb0014497},
   pages = {203--222},
   publisher = {{Springer}}
 }
