### 2023-05-10 5039b81
Mark cells above buildings but not touching building and not touching top of domain by -4.
