+++ src/cpp/include/Protobuf.h
@@ -1,256 +1,0 @@
-// Copyright (C) 2022 Dag Wästberg
-// Licensed under the MIT License
-//
-//  Modified by Anders Logg 2022
-
-#ifndef DTCC_PROTOBUF
-#define DTCC_PROTOBUF
-
-#include "dtcc_model/PointCloudMethods.h"
-#include "dtcc_model/PolygonMethods.h"
-#include "dtcc_model/dtcc.pb.h"
-
-#include "Logging.h"
-#include "Point.h"
-#include "PointCloud.h"
-#include "PointCloudProcessor.h"
-#include "Surface.h"
-
-#include "datamodel/Building.h"
-#include "datamodel/CityModel.h"
-
-namespace DTCC_BUILDER
-{
-
-class Protobuf
-{
-public:
-  // Developer note: Started but not completed migration to read/write functions
-  // (issue #52) Developer note: DTCC Builder object / DTCC Protobuf _object
-
-  // Read CityModel from Protobuf string
-  static void read(CityModel &city_model, const std::string &pb)
-  {
-    DTCC::CityModel _city_model{};
-    _city_model.ParseFromString(pb);
-    read(city_model, _city_model);
-  }
-
-  // Read CityModel from Protobuf object
-  static void read(CityModel &city_model, const DTCC::CityModel &_city_model)
-  {
-    for (const auto &_building : _city_model.buildings())
-    {
-      Building building{};
-      building.UUID = _building.uuid();
-      for (const auto &_vertex : _building.footprint().shell().vertices())
-        building.Footprint.Vertices.push_back(
-            Point2D(_vertex.x(), _vertex.y()));
-      city_model.Buildings.push_back(building);
-    }
-  }
-
-  static void write(const CityModel &cityModel, std::string &pb)
-  {
-    DTCC::CityModel pb_cm;
-
-    std::vector<DTCC::Building> pb_buildings;
-    info("Serializing " + str(cityModel.Buildings.size()) + " buildings");
-    for (const auto &building : cityModel.Buildings)
-    {
-      DTCC::Building pb_bld;
-      pb_bld.set_height(building.Height);
-      pb_bld.set_uuid(building.UUID);
-      pb_bld.set_error(building.error);
-      pb_bld.set_groundheight(building.GroundHeight);
-
-      std::vector<DTCC::Vector2D> pb_verts;
-      for (const auto &vert : building.Footprint.Vertices)
-      {
-        pb_verts.push_back(DTCC::Vertex(vert.x, vert.y));
-      }
-
-      pb_bld.mutable_footprint()->CopyFrom(CreatePolygon(pb_verts));
-
-      std::vector<DTCC::Vector3D> pb_pts;
-      for (const auto &pt : building.RoofPoints)
-      {
-        pb_pts.push_back(DTCC::Point(pt.x, pt.y, pt.z));
-      }
-      auto pb_roofpoints = DTCC::CreatePointCloud(pb_pts);
-      pb_bld.mutable_roofpoints()->Swap(&pb_roofpoints);
-
-      pb_buildings.push_back(pb_bld);
-    }
-
-    google::protobuf::RepeatedPtrField<DTCC::Building> building_data(
-        pb_buildings.begin(), pb_buildings.end());
-
-    pb_cm.mutable_buildings()->Swap(&building_data);
-
-    pb_cm.mutable_georeference()->set_x0(cityModel.Origin.x);
-    pb_cm.mutable_georeference()->set_y0(cityModel.Origin.y);
-
-    pb_cm.SerializeToString(&pb);
-  }
-
-  // FIXME: Rename to void read()
-  static PointCloud LoadPointCloud(std::string protobuf_string)
-  {
[…]
