### 2023-05-16 5b75ec4
Add some debugging
