+++ include/MeshGenerator.h
@@ -218,6 +218,9 @@ public:
     std::vector<short int> numCellsKept(mesh2D.Cells.size());
     std::fill(numCellsKept.begin(), numCellsKept.end(), 0);
 
+    // Keep track of buildings that are too low (will be removed)
+    std::set<size_t> lowBuildings;
+
     // Phase 1: Mark which cells that should be kept. This requires some
     // care since we want to mark all cells in a layer that belong to
     // the same building in the same way.
@@ -249,6 +252,8 @@ public:
             if (p.z < cityModel.Buildings[marker].MaxHeight())
             {
               keepBuilding[marker] = false;
+              if (layer == 0)
+                lowBuildings.insert(marker);
               break;
             }
           }
@@ -275,7 +280,9 @@ public:
 
             // Mark cell for removal (below building height)
             if (!keepBuilding[marker])
+            {
               keepCell[cellIndex] = false;
+            }
             else
             {
               // Mark cells that are above a building but not touching the
@@ -292,6 +299,13 @@ public:
         }
       }
     }
+
+    // Print warning for low buildings
+    for (const auto &marker : lowBuildings)
+    {
+      warning("MeshGenerator: Building " + str(marker) + " is lower than first layer, will be removed");
+    }
+
 
     // Phase 2: Extract new mesh for all cells that should be kept
 
@@ -323,8 +337,8 @@ public:
     }
 
     // Print vertex map (for debugging)
-    for (const auto v : vertexMap)
-      std::cout << "Vertex map: " << v.first << " --> " << v.second << std::endl;
+    //for (const auto v : vertexMap)
+    //  std::cout << "Vertex map: " << v.first << " --> " << v.second << std::endl;
 
     // Initialize new mesh data
     const size_t numVertices = vertexMap.size();
