### 2023-05-16 6d1c098
Fix warning message for low buildings
