### 2023-05-16 89611f0
Mark low buildings as halos, still not working properly. One building remains.
