### 2023-05-17 6aa2658
Add some more debugging
