### 2023-05-17 645567f
Rework mesh trimming and marking. Should be in good shape now.
