### 2023-05-20 #4928 @crvernon
@anderslogg @vbassn ... There are still formatting issues with the reference in your paper.

**Lines 25-29:**

> Previous research on 3D mesh generation for cities demonstrated, but do not fully account for the issues at hand (Ledoux et al., 2019),[Ledoux et al. (2021),(Paden et al., 2022). DTCC Builder aims to solve these challenges by automating the mesh generation process in a both robust and efficient way. Related work on 3D mesh generation for city modeling include (Ledoux et al., 2019), (Ledoux et al., 2021), and (Paden et al., 2022).

should simply be formatted as (and dropping the last sentence):

> Previous research on 3D mesh generation for cities demonstrated, but do not fully account for the issues at hand (Ledoux et al., 2019; Ledoux et al. 2021; Paden et al., 2022). DTCC Builder aims to solve these challenges by automating the mesh generation process in a both robust and efficient way.

To do this, the proper formatting in markdown would be:

> Previous research on 3D mesh generation for cities demonstrated, but do not fully account for the issues at hand [@ledouxCityJSONCompactEasytouse2019; @ledoux3dfierAutomaticReconstruction2021; @padenAutomaticReconstruction3D2022]. DTCC Builder aims to solve these challenges by automating the mesh generation process in a both robust and efficient way.

**Line 44:**

> smoothing process as described by (Naserentin & Logg, 2022).

should be formatted as:

> smoothing process as described by Naserentin & Logg (2022).

To do this, the proper formatting in markdown would be:

>smoothing process as described by @naserentinDigitalTwinsCity2022.
