### 2023-05-20 b8cedc0
Updated according to https://github.com/openjournals/joss-reviews/issues/4928#issuecomment-1555911960
