### 2023-05-20 e854652
Updated paper.md according to https://github.com/openjournals/joss-reviews/issues/4928#issuecomment-1555911960
