### 2023-05-23 b05fddb
cmake cleanup
