### 2023-05-23 06b159e
remove external eigen dependency
