+++ CMakeLists.txt
@@ -29,7 +29,7 @@ include(cmake/FindLibGEOS.cmake)
 find_package(Triangle REQUIRED)
 find_package(nlohmann_json REQUIRED)
 find_package(LibGEOS 3.16 REQUIRED)
-find_package(Eigen3 REQUIRED)
+# find_package(Eigen3 REQUIRED)
 
 # Find pybind11
 set(pybind11_DIR /usr/lib/cmake/pybind11)
@@ -59,7 +59,6 @@ include_directories(${CMAKE_SOURCE_DIR}/src/cpp/include)
 include_directories(${CMAKE_SOURCE_DIR}/src/cpp/include/citymodel)
 include_directories(${CMAKE_SOURCE_DIR}/src/cpp/external)
 include_directories(/usr/include/nlohmann/)
-include_directories(/usr/include/eigen3/)
 include_directories(/usr/local/include)
 
 # Set installation directory
