+++ CMakeLists.txt
@@ -27,9 +27,7 @@ include(cmake/FindLibGEOS.cmake)
 
 # Find packages
 find_package(Triangle REQUIRED)
-find_package(nlohmann_json REQUIRED)
 find_package(LibGEOS 3.16 REQUIRED)
-# find_package(Eigen3 REQUIRED)
 
 # Find pybind11
 set(pybind11_DIR /usr/lib/cmake/pybind11)
@@ -58,7 +56,6 @@ endif()
 include_directories(${CMAKE_SOURCE_DIR}/src/cpp/include)
 include_directories(${CMAKE_SOURCE_DIR}/src/cpp/include/citymodel)
 include_directories(${CMAKE_SOURCE_DIR}/src/cpp/external)
-include_directories(/usr/include/nlohmann/)
 include_directories(/usr/local/include)
 
 # Set installation directory
