### 2023-05-26 5ed76e0
move building height to processors
