+++ src/cpp/include/MeshGenerator.h
@@ -34,10 +34,13 @@ public:
   // defined by (xMin, xMax) x (yMin, yMax). The edges of the mesh respect the
   // boundaries of the buildings.
   //
-  // The domain markers label the triangles inside building footprints
-  // with the number of the building (0, 1, 2, ...). Triangles that are
-  // neighbors of building triangles (just outside buildings) are marked
-  // as -1 and the remaining triangles (the ground) are marked as -2.
+  // Markers:
+  //
+  // -2: ground (cells outside buildings and halos)
+  // -1: halos (cells close to buildings)
+  //  0: building 0 (cells inside building 0)
+  //  1: building 1 (cells inside building 1)
+  //  etc (non-negative integers mark cells inside buildings)
   static void GenerateMesh2D(Mesh2D &mesh2D,
                              const CityModel &cityModel,
                              const BoundingBox2D &boundingBox,
@@ -78,9 +81,17 @@ public:
   // Generate 3D mesh. The mesh is a tetrahedral mesh generated by
   // extruding the 2D mesh in the vertical (z) direction.
   //
-  // The domain markers for the 3D mesh are propagated as is from
-  // the domain markers of the 2D mesh, except for cells in the top
-  // layer that are marked as -3.
+  // Markers:
+  //
+  // -2: ground (cells outside buildings and halos)
+  // -1: halos (cells close to buildings)
+  //  0: building 0 (cells inside building 0)
+  //  1: building 1 (cells inside building 1)
+  //  etc (non-negative integers mark cells inside buildings)
+  //
+  // Note that the markers are just propagatated upward from the
+  // 2D mesh, meaning that the markers will be the same in each
+  // column of the 3D mesh.
   //
   // It is assumed that the 2D mesh is sorted, which is the case if the
   // mesh has been generated by calling GenerateMesh2D().
@@ -93,6 +104,8 @@ public:
   {
     info("MeshGenerator: Generating 3D mesh...");
     Timer timer("GenerateMesh3D");
+
+    info("CHECK");
 
     // Compute number of layers
     const size_t numLayers = int(std::ceil(domainHeight / meshResolution));
@@ -160,7 +173,29 @@ public:
       {
         for (const auto &marker : mesh2D.Markers)
         {
-          const int m = (layer == numLayers - 1 ? -3 : marker);
+          int m = 0;
+
+          // Top layer marked as -3
+          if (layer == numLayers - 1)
+            m = -3;
+
+          // Halo and ground only marked for bottom layer
+          else if (marker == -1 || marker == -2)
+          {
+            if (layer == 0)
+              m = marker;
+            else
+              m = -4;
+          }
+
+          // Buildings marked for all layers (except top layer).
+          // Later adjusted to -4 above buildings in TrimMesh3D.
+          else
+          {
+            m = marker;
+          }
+
+          // Set markers
           mesh3D.Markers[k++] = m;
           mesh3D.Markers[k++] = m;
           mesh3D.Markers[k++] = m;
@@ -172,7 +207,24 @@ public:
   }
 
   // Trim 3D mesh. The mesh is trimmed by removing tetrahedra inside
-  // building shape. The domain markers are left untouched.
+  // building shape.
+  //
+  // Markers:
+  //
+  // -4: other (not top, ground, halo, or building)
+  // -3: top (cells in *top layer*)
+  // -2: ground (cells in *bottom layer* outside buildings and halos)
+  // -1: halos (cells in *bottom layer* close to buildings)
+  //  0: building 0 (cells *first layer* inside building 0)
[…]
